% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-adipot.R, R/cdo-adisit.R, R/family-adisit.R
\name{cdo_adipot}
\alias{cdo_adipot}
\alias{cdo_adisit}
\alias{adisit}
\title{Potential temperature to insitu temperature and vice versa}
\usage{
cdo_adipot(ifile, pressure = NULL, ofile = NULL)

cdo_adisit(ifile, pressure = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{pressure, }{FLOAT - Pressure in bar (constant value assigned to all levels)}

\item{ofile}{String with the path to the output file.}
}
\description{
Potential temperature to insitu temperature and vice versa
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{adisit  Potential temperature to in-situ temperature
        This is a special operator for the post processing of the ocean and sea ice model MPIOM.
        It converts potential temperature adiabatically to in-situ temperature to(t, s, p).
        Required input fields are sea water potential temperature (name=tho; code=2) and sea water salinity (name=sao; code=5).
        Pressure is calculated from the level information or can be specified by the optional parameter.
        Output fields are sea water temperature (name=to; code=20) and sea water salinity (name=s; code=5).
adipot  In-situ temperature to potential temperature
        This is a special operator for the post processing of the ocean and sea ice model MPIOM.
        It converts in-situ temperature to potential temperature tho(to, s, p). Required input fields
        are sea water in-situ temperature (name=t; code=2) and sea water salinity (name=sao,s; code=5).
        Pressure is calculated from the level information or can be specified by the optional parameter.
        Output fields are sea water temperature (name=tho; code=2) and sea water salinity (name=s; code=5).
}\if{html}{\out{</div>}}
}
