% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-after.R, R/family-afterburner.R
\name{cdo_after}
\alias{cdo_after}
\alias{afterburner}
\title{ECHAM standard post processor}
\usage{
cdo_after(ifiles, vct = NULL, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{vct}{STRING - File with VCT in ASCII format}

\item{ofile}{String with the path to the output file.}
}
\description{
The \"afterburner\" is the standard post processor for ECHAM GRIB and NetCDF data which provides the following operations:  - Extract specified variables and levels - Compute derived variables - Transform spectral data to Gaussian grid representation - Vertical interpolation to pressure levels - Compute temporal means  This operator reads selection parameters as namelist from stdin. Use the UNIX redirection \"<namelistfile\" to read the namelist from file.  The input files can't be combined with other CDO operators because of an optimized reader for this operator.
}
