% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-divdpm.R, R/cdo-divdpy.R, R/cdo-muldpm.R,
%   R/cdo-muldpy.R, R/family-arithdays.R
\name{cdo_divdpm}
\alias{cdo_divdpm}
\alias{cdo_divdpy}
\alias{cdo_muldpm}
\alias{cdo_muldpy}
\alias{arithdays}
\title{Arithmetic with days}
\usage{
cdo_divdpm(ifile, ofile = NULL)

cdo_divdpy(ifile, ofile = NULL)

cdo_muldpm(ifile, ofile = NULL)

cdo_muldpy(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
This module multiplies or divides each timestep of a dataset with the corresponding days per month or days per year. The result of these functions depends on the used calendar of the input data.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{muldpm  Multiply with days per month
        o(t,x) = i(t,x) * days_per_month
divdpm  Divide by days per month
        o(t,x) = i(t,x) / days_per_month
muldpy  Multiply with days per year
        o(t,x) = i(t,x) * days_per_year
divdpy  Divide by days per year
        o(t,x) = i(t,x) / days_per_year
}\if{html}{\out{</div>}}
}
