% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-divcoslat.R, R/cdo-mulcoslat.R,
%   R/family-arithlat.R
\name{cdo_divcoslat}
\alias{cdo_divcoslat}
\alias{cdo_mulcoslat}
\alias{arithlat}
\title{Arithmetic with latitude}
\usage{
cdo_divcoslat(ifile, ofile = NULL)

cdo_mulcoslat(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
This module multiplies or divides each field element with the cosine of the latitude.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{mulcoslat  Multiply with the cosine of the latitude
           o(t,x) = i(t,x) * cos(latitude(x))
divcoslat  Divide by cosine of the latitude
           o(t,x) = i(t,x) / cos(latitude(x))
}\if{html}{\out{</div>}}
}
