% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.R
\name{cdo}
\alias{cdo}
\alias{cdo_operator}
\title{Execute a custom operator}
\usage{
cdo(operator, input, params = NULL, output = NULL)

cdo_operator(command, params, n_input, n_output)
}
\arguments{
\item{operator, }{a list created with \code{cdo_operator}.}

\item{input}{a list with the input files.}

\item{params}{a character vector with the name of the parametsr}

\item{output}{a vector of file name(s).}

\item{command}{a string with the command used to run the operator}

\item{n_input, n_output}{an integer with the number of input and ouput files
required by the operator}
}
\value{
a cdo operation.

A list with elements comand, params, n_input and n_output.
}
\description{
Execute a custom operator
}
