% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-abs.R
\name{cdo_abs}
\alias{cdo_abs}
\title{abs}
\usage{
cdo_abs(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
