% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-acos.R
\name{cdo_acos}
\alias{cdo_acos}
\title{acos}
\usage{
cdo_acos(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
