% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-after.R
\name{cdo_after}
\alias{cdo_after}
\title{after}
\usage{
cdo_after(..., vct = NULL, ofile = NULL)
}
\arguments{
\item{vct, }{STRING - File with VCT in ASCII format}

\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
ECHAM standard post processor
}
