% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ap2pl.R
\name{cdo_ap2pl}
\alias{cdo_ap2pl}
\title{ap2pl}
\usage{
cdo_ap2pl(ifile, plevels = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{plevels, }{FLOAT - Comma-separated list of pressure levels in pascal}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Vertical pressure interpolation
}
