% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-apply.R
\name{cdo_apply}
\alias{cdo_apply}
\title{apply}
\usage{
cdo_apply(ifile, operators = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{operators, }{STRING - Blank-separated list of CDO operators.}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Apply operators
}
