% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-asin.R
\name{cdo_asin}
\alias{cdo_asin}
\title{asin}
\usage{
cdo_asin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
