% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-atan.R
\name{cdo_atan}
\alias{cdo_atan}
\title{atan}
\usage{
cdo_atan(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
