% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-bandpass.R
\name{cdo_bandpass}
\alias{cdo_bandpass}
\title{bandpass}
\usage{
cdo_bandpass(ifile, fmin = NULL, fmax = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{fmin, }{FLOAT    Minimum - frequency per year that passes the filter.}

\item{fmax, }{FLOAT    Maximum - frequency per year that passes the filter.}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Time series filtering
}
