% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-bottomvalue.R
\name{cdo_bottomvalue}
\alias{cdo_bottomvalue}
\title{bottomvalue}
\usage{
cdo_bottomvalue(ifile, isovalue = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{isovalue, }{FLOAT - Isosurface value}

\item{ofile}{String with the path to the output file.}
}
\description{
Extract surface
}
