% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-cat.R
\name{cdo_cat}
\alias{cdo_cat}
\title{cat}
\usage{
cdo_cat(..., ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Copy datasets
}
