% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-changemulti.R
\name{cdo_changemulti}
\alias{cdo_changemulti}
\title{changemulti}
\usage{
cdo_changemulti(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Select multiple fields via GRIB1 parameters
}
