% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-cmorlite.R
\name{cdo_cmorlite}
\alias{cdo_cmorlite}
\title{cmorlite}
\usage{
cdo_cmorlite(ifile, table = NULL, convert = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{table, }{STRING - Name of the CMOR table as specified from PCMDI}

\item{convert, }{STRING - Converts the units if necessary}

\item{ofile}{String with the path to the output file.}
}
\description{
CMOR lite
}
