% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-collgrid.R
\name{cdo_collgrid}
\alias{cdo_collgrid}
\title{collgrid}
\usage{
cdo_collgrid(..., nx = NULL, names = NULL, ofile = NULL)
}
\arguments{
\item{nx, }{INTEGER - Number of regions in x direction \link{default: number of input files}}

\item{names, }{STRING - Comma-separated list of variable names \link{default: all variables}}

\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Collect horizontal grid
}
