% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-utilize.R
\name{cdo_combine}
\alias{cdo_combine}
\alias{cdo_combine_large}
\title{cdo_combine}
\usage{
cdo_combine(
  files,
  outfile = NULL,
  outdir = ".",
  range = c(70, 140, 15, 55),
  delta = 3,
  cdo = "cdo",
  ncluster = 4,
  verbose = TRUE,
  run = FALSE,
  ...
)

cdo_combine_large(files, outfile = NULL, chunk = 100, ...)
}
\description{
cdo_combine
}
