% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-consecsum.R
\name{cdo_consecsum}
\alias{cdo_consecsum}
\title{consecsum}
\usage{
cdo_consecsum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Consecute timestep periods
}
