% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-consects.R
\name{cdo_consects}
\alias{cdo_consects}
\title{consects}
\usage{
cdo_consects(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Consecute timestep periods
}
