% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-const.R
\name{cdo_const}
\alias{cdo_const}
\title{const}
\usage{
cdo_const(
  ifile,
  const = NULL,
  seed = NULL,
  grid = NULL,
  start = NULL,
  end = NULL,
  inc = NULL,
  levels = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{Strings with the path to the input files.}

\item{const, }{FLOAT - Constant}

\item{seed, }{INTEGER - The seed for a new sequence of pseudo-random numbers [default: 1]}

\item{grid, }{STRING - Target grid description file or name}

\item{start, }{FLOAT - Start value of the loop}

\item{end, }{FLOAT - End value of the loop}

\item{inc, }{FLOAT - Increment of the loop [default: 1]}

\item{levels, }{FLOAT - Target levels in metre above surface}

\item{ofile}{String with the path to the output file.}
}
\description{
Generate a field
}
