% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-contour.R
\name{cdo_contour}
\alias{cdo_contour}
\title{contour}
\usage{
cdo_contour(ifile, parameter = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{parameter, }{STRING - Comma-separated list of plot parameters}

\item{ofile}{String with the path to the output file.}
}
\description{
Lat/Lon plot
}
