% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-copy.R
\name{cdo_copy}
\alias{cdo_copy}
\title{copy}
\usage{
cdo_copy(ifiles, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Copy datasets
}
