% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-cos.R
\name{cdo_cos}
\alias{cdo_cos}
\title{cos}
\usage{
cdo_cos(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
