% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dayadd.R
\name{cdo_dayadd}
\alias{cdo_dayadd}
\title{dayadd}
\usage{
cdo_dayadd(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Daily arithmetic
}
