% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dayavg.R
\name{cdo_dayavg}
\alias{cdo_dayavg}
\title{dayavg}
\usage{
cdo_dayavg(ifile, complete_only = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{complete_only, }{BOOL - Process only complete years}

\item{ofile}{String with the path to the output file.}
}
\description{
Daily statistics
}
