% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-daydiv.R
\name{cdo_daydiv}
\alias{cdo_daydiv}
\title{daydiv}
\usage{
cdo_daydiv(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Daily arithmetic
}
