% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-daymax.R
\name{cdo_daymax}
\alias{cdo_daymax}
\title{daymax}
\usage{
cdo_daymax(ifile, complete_only = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{complete_only, }{BOOL - Process only complete years}

\item{ofile}{String with the path to the output file.}
}
\description{
Daily statistics
}
