% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-daymul.R
\name{cdo_daymul}
\alias{cdo_daymul}
\title{daymul}
\usage{
cdo_daymul(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Daily arithmetic
}
