% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-daypctl.R
\name{cdo_daypctl}
\alias{cdo_daypctl}
\title{daypctl}
\usage{
cdo_daypctl(ifile, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{p, }{FLOAT - Percentile number in {0, ..., 100}}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Daily percentile values
}
