% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-daysum.R
\name{cdo_daysum}
\alias{cdo_daysum}
\title{daysum}
\usage{
cdo_daysum(ifile, complete_only = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{complete_only, }{BOOL - Process only complete years}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Daily statistics
}
