% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-delgridcell.R
\name{cdo_delgridcell}
\alias{cdo_delgridcell}
\title{delgridcell}
\usage{
cdo_delgridcell(ifile, indices = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{indices, }{INTEGER - Comma-separated list or first/last\link{/inc} range of indices}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Select grid cells
}
