% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-delmulti.R
\name{cdo_delmulti}
\alias{cdo_delmulti}
\title{delmulti}
\usage{
cdo_delmulti(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Select multiple fields via GRIB1 parameters
}
