% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-detrend.R
\name{cdo_detrend}
\alias{cdo_detrend}
\title{detrend}
\usage{
cdo_detrend(ifile, equal = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{equal, }{BOOL - Set to false for unequal distributed timesteps (default: true)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Detrend time series
}
