% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhourmax.R
\name{cdo_dhourmax}
\alias{cdo_dhourmax}
\title{dhourmax}
\usage{
cdo_dhourmax(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
