% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhourmean.R
\name{cdo_dhourmean}
\alias{cdo_dhourmean}
\title{dhourmean}
\usage{
cdo_dhourmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
