% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhourrange.R
\name{cdo_dhourrange}
\alias{cdo_dhourrange}
\title{dhourrange}
\usage{
cdo_dhourrange(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
