% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhourstd.R
\name{cdo_dhourstd}
\alias{cdo_dhourstd}
\title{dhourstd}
\usage{
cdo_dhourstd(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
