% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhourstd1.R
\name{cdo_dhourstd1}
\alias{cdo_dhourstd1}
\title{dhourstd1}
\usage{
cdo_dhourstd1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
