% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhoursum.R
\name{cdo_dhoursum}
\alias{cdo_dhoursum}
\title{dhoursum}
\usage{
cdo_dhoursum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
