% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhourvar1.R
\name{cdo_dhourvar1}
\alias{cdo_dhourvar1}
\title{dhourvar1}
\usage{
cdo_dhourvar1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
