% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-distgrid.R
\name{cdo_distgrid}
\alias{cdo_distgrid}
\title{distgrid}
\usage{
cdo_distgrid(ifile, nx = NULL, ny = NULL, obase = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{nx, }{INTEGER - Number of regions in x direction, or number of pieces for unstructured grids}

\item{ny, }{INTEGER - Number of regions in y direction \link{default: 1}}

\item{obase}{string with the basename of the output files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Distribute horizontal grid
}
