% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-div.R
\name{cdo_div}
\alias{cdo_div}
\title{div}
\usage{
cdo_div(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic on two datasets
}
