% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-divcoslat.R
\name{cdo_divcoslat}
\alias{cdo_divcoslat}
\title{divcoslat}
\usage{
cdo_divcoslat(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic with latitude
}
