% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-divdpm.R
\name{cdo_divdpm}
\alias{cdo_divdpm}
\title{divdpm}
\usage{
cdo_divdpm(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic with days
}
