% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dv2ps.R
\name{cdo_dv2ps}
\alias{cdo_dv2ps}
\title{dv2ps}
\usage{
cdo_dv2ps(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
D and V to velocity potential and stream function
}
