% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_cdd.R
\name{cdo_eca_cdd}
\alias{cdo_eca_cdd}
\title{eca_cdd}
\usage{
cdo_eca_cdd(ifile, R = NULL, N = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{R, }{FLOAT - Precipitation threshold (unit: mm; default: R = 1 mm)}

\item{N, }{INTEGER - Minimum number of days exceeded (default: N = 5)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Consecutive dry days index per time period
}
