% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_cfd.R
\name{cdo_eca_cfd}
\alias{cdo_eca_cfd}
\title{eca_cfd}
\usage{
cdo_eca_cfd(ifile, N = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{N, }{INTEGER - Minimum number of days exceeded (default: N = 5)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Consecutive frost days index per time period
}
