% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_csu.R
\name{cdo_eca_csu}
\alias{cdo_eca_csu}
\title{eca_csu}
\usage{
cdo_eca_csu(ifile, T = NULL, N = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{T, }{FLOAT - Temperature threshold (unit: °C; default: T = 25°C)}

\item{N, }{INTEGER - Minimum number of days exceeded (default: N = 5)}

\item{ofile}{String with the path to the output file.}
}
\description{
Consecutive summer days index per time period
}
