% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_cwfi.R
\name{cdo_eca_cwfi}
\alias{cdo_eca_cwfi}
\title{eca_cwfi}
\usage{
cdo_eca_cwfi(ifile1, ifile2, nday = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{nday, }{INTEGER - Number of consecutive days (default: nday = 6)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Cold
}
