% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_etr.R
\name{cdo_eca_etr}
\alias{cdo_eca_etr}
\title{eca_etr}
\usage{
cdo_eca_etr(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Intra
}
