% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_hwdi.R
\name{cdo_eca_hwdi}
\alias{cdo_eca_hwdi}
\title{eca_hwdi}
\usage{
cdo_eca_hwdi(ifile1, ifile2, nday = NULL, T = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{nday, }{INTEGER - Number of consecutive days (default: nday = 6)}

\item{T, }{FLOAT - Temperature offset (unit: °C; default: T = 5°C)}

\item{ofile}{String with the path to the output file.}
}
\description{
Heat wave duration index wrt mean of reference period
}
