% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_id.R
\name{cdo_eca_id}
\alias{cdo_eca_id}
\title{eca_id}
\usage{
cdo_eca_id(ifile, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Ice days index per time period
}
