% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_r95p.R
\name{cdo_eca_r95p}
\alias{cdo_eca_r95p}
\title{eca_r95p}
\usage{
cdo_eca_r95p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Very wet days wrt 95th percentile of reference period
}
