% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_r99p.R
\name{cdo_eca_r99p}
\alias{cdo_eca_r99p}
\title{eca_r99p}
\usage{
cdo_eca_r99p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Extremely wet days wrt 99th percentile of reference period
}
