% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_rr1.R
\name{cdo_eca_rr1}
\alias{cdo_eca_rr1}
\title{eca_rr1}
\usage{
cdo_eca_rr1(ifile, R = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{R, }{FLOAT - Precipitation threshold (unit: mm; default: R = 1 mm)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Wet days index per time period
}
