% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_rx5day.R
\name{cdo_eca_rx5day}
\alias{cdo_eca_rx5day}
\title{eca_rx5day}
\usage{
cdo_eca_rx5day(ifile, x = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{x, }{FLOAT - Precipitation threshold (unit: mm; default: x = 50 mm)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Highest five
}
