% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_su.R
\name{cdo_eca_su}
\alias{cdo_eca_su}
\title{eca_su}
\usage{
cdo_eca_su(ifile, T = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{T, }{FLOAT - Temperature threshold (unit: °C; default: T = 25°C)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Summer days index per time period
}
