% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_tg90p.R
\name{cdo_eca_tg90p}
\alias{cdo_eca_tg90p}
\title{eca_tg90p}
\usage{
cdo_eca_tg90p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Warm days percent wrt 90th percentile of reference period
}
