% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_tn10p.R
\name{cdo_eca_tn10p}
\alias{cdo_eca_tn10p}
\title{eca_tn10p}
\usage{
cdo_eca_tn10p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Cold nights percent wrt 10th percentile of reference period
}
