% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_tr.R
\name{cdo_eca_tr}
\alias{cdo_eca_tr}
\title{eca_tr}
\usage{
cdo_eca_tr(ifile, T = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{T, }{FLOAT - Temperature threshold (unit: °C; default: T = 20°C)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Tropical nights index per time period
}
