% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-enlarge.R
\name{cdo_enlarge}
\alias{cdo_enlarge}
\title{enlarge}
\usage{
cdo_enlarge(ifile, grid = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{grid, }{STRING - Target grid description file or name}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Enlarge fields
}
