% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensbrs.R
\name{cdo_ensbrs}
\alias{cdo_ensbrs}
\title{ensbrs}
\usage{
cdo_ensbrs(..., obase = NULL)
}
\arguments{
\item{obase}{string with the basename of the output files.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}
}
\description{
Ensemble validation tools
}
