% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-enscrps.R
\name{cdo_enscrps}
\alias{cdo_enscrps}
\title{enscrps}
\usage{
cdo_enscrps(..., obase = NULL)
}
\arguments{
\item{obase}{string with the basename of the output files.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}
}
\description{
Ensemble validation tools
}
