% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensmax.R
\name{cdo_ensmax}
\alias{cdo_ensmax}
\title{ensmax}
\usage{
cdo_ensmax(ifiles, p = NULL, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over an ensemble
}
