% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensmean.R
\name{cdo_ensmean}
\alias{cdo_ensmean}
\title{ensmean}
\usage{
cdo_ensmean(ifiles, p = NULL, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over an ensemble
}
