% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensrkhistspace.R
\name{cdo_ensrkhistspace}
\alias{cdo_ensrkhistspace}
\title{ensrkhistspace}
\usage{
cdo_ensrkhistspace(..., ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Statistical values over an ensemble
}
