% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensrkhisttime.R
\name{cdo_ensrkhisttime}
\alias{cdo_ensrkhisttime}
\title{ensrkhisttime}
\usage{
cdo_ensrkhisttime(ifiles, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over an ensemble
}
