% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensroc.R
\name{cdo_ensroc}
\alias{cdo_ensroc}
\title{ensroc}
\usage{
cdo_ensroc(ifiles, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over an ensemble
}
