% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eof3d.R
\name{cdo_eof3d}
\alias{cdo_eof3d}
\title{eof3d}
\usage{
cdo_eof3d(ifile, neof = NULL, ofile1 = NULL, ofile2 = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{neof, }{INTEGER - Number of eigen functions}

\item{ofile1, ofile2}{Strings with the path to the output files.}
}
\description{
Empirical Orthogonal Functions
}
