% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eofcoeff.R
\name{cdo_eofcoeff}
\alias{cdo_eofcoeff}
\title{eofcoeff}
\usage{
cdo_eofcoeff(ifile1, ifile2, obase = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{obase}{String with the basename of the output files.}
}
\description{
Principal coefficients of EOFs
}
