% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eofspatial.R
\name{cdo_eofspatial}
\alias{cdo_eofspatial}
\title{eofspatial}
\usage{
cdo_eofspatial(ifile, neof = NULL, ofile1 = NULL, ofile2 = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{neof, }{INTEGER - Number of eigen functions}

\item{ofile1, ofile2}{Strings with the path to the output files.}
}
\description{
Empirical Orthogonal Functions
}
