% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eq.R
\name{cdo_eq}
\alias{cdo_eq}
\title{eq}
\usage{
cdo_eq(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
compares two datasets field by field
}
