% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eqc.R
\name{cdo_eqc}
\alias{cdo_eqc}
\title{eqc}
\usage{
cdo_eqc(ifile, c, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{c, }{Constant}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Compares all fields of a dataset with a constant
}
