% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-etccdi_csdi.R
\name{cdo_etccdi_csdi}
\alias{cdo_etccdi_csdi}
\title{etccdi_csdi}
\usage{
cdo_etccdi_csdi(ifile1, ifile2, nday = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{nday, }{INTEGER - Number of consecutive days (default: nday = 6)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Cold
}
