% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-etccdi_id.R
\name{cdo_etccdi_id}
\alias{cdo_etccdi_id}
\title{etccdi_id}
\usage{
cdo_etccdi_id(ifile, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Ice days index per time period
}
