% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-etccdi_r1mm.R
\name{cdo_etccdi_r1mm}
\alias{cdo_etccdi_r1mm}
\title{etccdi_r1mm}
\usage{
cdo_etccdi_r1mm(ifile, x = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{x, }{FLOAT - Daily precipitation amount threshold in [mm]}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Precipitation days index per time period
}
