% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-etccdi_rx1day.R
\name{cdo_etccdi_rx1day}
\alias{cdo_etccdi_rx1day}
\title{etccdi_rx1day}
\usage{
cdo_etccdi_rx1day(ifile, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Highest one day precipitation amount per time period
}
