% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-etccdi_rx5day.R
\name{cdo_etccdi_rx5day}
\alias{cdo_etccdi_rx5day}
\title{etccdi_rx5day}
\usage{
cdo_etccdi_rx5day(ifile, x = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{x, }{FLOAT - Precipitation threshold (unit: mm; default: x = 50 mm)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Highest five
}
