% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-etccdi_tr.R
\name{cdo_etccdi_tr}
\alias{cdo_etccdi_tr}
\title{etccdi_tr}
\usage{
cdo_etccdi_tr(ifile, T = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{T, }{FLOAT - Temperature threshold (unit: °C; default: T = 20°C)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Tropical nights index per time period
}
