% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-exp.R
\name{cdo_exp}
\alias{cdo_exp}
\title{exp}
\usage{
cdo_exp(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
