% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-expr.R
\name{cdo_expr}
\alias{cdo_expr}
\title{expr}
\usage{
cdo_expr(ifile, instr = NULL, filename = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{instr, }{STRING - Processing instructions (need to be 'quoted' in most cases)}

\item{filename, }{STRING - File with processing instructions}

\item{ofile}{String with the path to the output file.}
}
\description{
Evaluate expressions
}
