% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-exprf.R
\name{cdo_exprf}
\alias{cdo_exprf}
\title{exprf}
\usage{
cdo_exprf(ifile, instr = NULL, filename = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{instr, }{STRING - Processing instructions (need to be 'quoted' in most cases)}

\item{filename, }{STRING - File with processing instructions}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Evaluate expressions
}
