% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fdns.R
\name{cdo_fdns}
\alias{cdo_fdns}
\title{fdns}
\usage{
cdo_fdns(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Frost days where no snow index per time period
}
