% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fldavg.R
\name{cdo_fldavg}
\alias{cdo_fldavg}
\title{fldavg}
\usage{
cdo_fldavg(ifile, weights = NULL, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{weights, }{BOOL - weights=FALSE disables weighting by grid cell area \link{default: weights=TRUE}}

\item{p, }{FLOAT - Percentile number in {0, ..., 100}}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Statistical values over a field
}
