% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fldcor.R
\name{cdo_fldcor}
\alias{cdo_fldcor}
\title{fldcor}
\usage{
cdo_fldcor(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Correlation in grid space
}
