% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fldcovar.R
\name{cdo_fldcovar}
\alias{cdo_fldcovar}
\title{fldcovar}
\usage{
cdo_fldcovar(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Covariance in grid space
}
