% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fldkurt.R
\name{cdo_fldkurt}
\alias{cdo_fldkurt}
\title{fldkurt}
\usage{
cdo_fldkurt(ifile, weights = NULL, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{weights, }{BOOL - weights=FALSE disables weighting by grid cell area [default: weights=TRUE]}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over a field
}
