% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fourier.R
\name{cdo_fourier}
\alias{cdo_fourier}
\title{fourier}
\usage{
cdo_fourier(ifile, epsilon = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{epsilon, }{INTEGER - -1: forward transformation; 1: backward transformation}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Fourier transformation
}
