% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gencon.R
\name{cdo_gencon}
\alias{cdo_gencon}
\title{gencon}
\usage{
cdo_gencon(ifile, grid = NULL, map3d = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Target grid description file or name}

\item{map3d, }{BOOL - Generate all mapfiles of the first 3D field}

\item{ofile}{String with the path to the output file.}
}
\description{
First order conservative remapping
}
