% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gh2hl.R
\name{cdo_gh2hl}
\alias{cdo_gh2hl}
\title{gh2hl}
\usage{
cdo_gh2hl(ifile, hlevels = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{hlevels, }{FLOAT - Comma-separated list of height levels in meter}

\item{ofile}{String with the path to the output file.}
}
\description{
Vertical height interpolation
}
