% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gheight.R
\name{cdo_gheight}
\alias{cdo_gheight}
\title{gheight}
\usage{
cdo_gheight(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Derived model parameters
}
