% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gp2sp.R
\name{cdo_gp2sp}
\alias{cdo_gp2sp}
\title{gp2sp}
\usage{
cdo_gp2sp(ifile, type = NULL, trunc = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{type, }{STRING - Type of the grid: quadratic, linear, cubic (default: type=quadratic)}

\item{trunc, }{STRING - Triangular truncation}

\item{ofile}{String with the path to the output file.}
}
\description{
Spectral transformation
}
