% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-graph.R
\name{cdo_graph}
\alias{cdo_graph}
\title{graph}
\usage{
cdo_graph(..., parameter = NULL, ofile = NULL)
}
\arguments{
\item{parameter, }{STRING - Comma-separated list of plot parameters}

\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Line graph plot
}
