% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-utilize.R
\name{cdo_grid}
\alias{cdo_grid}
\title{get regular cdo grid}
\usage{
cdo_grid(
  range = c(70, 140, 15, 55),
  cellsize = 0.5,
  mid = FALSE,
  outfile = NULL
)
}
\arguments{
\item{range}{\verb{[lon_min, lon_max, lat_min, lat_max]}}

\item{cellsize}{double}

\item{mid}{If \code{TRUE}, the first coordinate is on the middle; otherwise,
on the edge.}

\item{outfile}{character, path of output file}
}
\description{
get regular cdo grid
}
