% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gridarea.R
\name{cdo_gridarea}
\alias{cdo_gridarea}
\title{gridarea}
\usage{
cdo_gridarea(ifile, radius = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{radius, }{FLOAT - Planet radius in meter}

\item{ofile}{String with the path to the output file.}
}
\description{
Grid cell quantities
}
