% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gridboxmax.R
\name{cdo_gridboxmax}
\alias{cdo_gridboxmax}
\title{gridboxmax}
\usage{
cdo_gridboxmax(ifile, nx = NULL, ny = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nx, }{INTEGER - Number of grid boxes in x direction}

\item{ny, }{INTEGER - Number of grid boxes in y direction}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over grid boxes
}
