% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gridweights.R
\name{cdo_gridweights}
\alias{cdo_gridweights}
\title{gridweights}
\usage{
cdo_gridweights(ifile, radius = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{radius, }{FLOAT - Planet radius in meter}

\item{ofile}{String with the path to the output file.}
}
\description{
Grid cell quantities
}
