% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gtc.R
\name{cdo_gtc}
\alias{cdo_gtc}
\title{gtc}
\usage{
cdo_gtc(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Comparison of a field with a constant
}
