% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-histmean.R
\name{cdo_histmean}
\alias{cdo_histmean}
\title{histmean}
\usage{
cdo_histmean(ifile, bounds = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{bounds, }{FLOAT - Comma-separated list of the bin bounds (-inf and inf valid)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Histogram
}
