% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-houravg.R
\name{cdo_houravg}
\alias{cdo_houravg}
\title{houravg}
\usage{
cdo_houravg(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
