% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hourmax.R
\name{cdo_hourmax}
\alias{cdo_hourmax}
\title{hourmax}
\usage{
cdo_hourmax(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
