% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hourmean.R
\name{cdo_hourmean}
\alias{cdo_hourmean}
\title{hourmean}
\usage{
cdo_hourmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
