% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hourmin.R
\name{cdo_hourmin}
\alias{cdo_hourmin}
\title{hourmin}
\usage{
cdo_hourmin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
