% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hourpctl.R
\name{cdo_hourpctl}
\alias{cdo_hourpctl}
\title{hourpctl}
\usage{
cdo_hourpctl(ifile, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{p, }{FLOAT - Percentile number in {0, ..., 100}}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Hourly percentile values
}
