% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hourrange.R
\name{cdo_hourrange}
\alias{cdo_hourrange}
\title{hourrange}
\usage{
cdo_hourrange(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
