% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hourstd.R
\name{cdo_hourstd}
\alias{cdo_hourstd}
\title{hourstd}
\usage{
cdo_hourstd(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
