% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hourstd1.R
\name{cdo_hourstd1}
\alias{cdo_hourstd1}
\title{hourstd1}
\usage{
cdo_hourstd1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
