% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hoursum.R
\name{cdo_hoursum}
\alias{cdo_hoursum}
\title{hoursum}
\usage{
cdo_hoursum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
