% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hourvar.R
\name{cdo_hourvar}
\alias{cdo_hourvar}
\title{hourvar}
\usage{
cdo_hourvar(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
