% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hourvar1.R
\name{cdo_hourvar1}
\alias{cdo_hourvar1}
\title{hourvar1}
\usage{
cdo_hourvar1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hourly statistics
}
