% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-hurr.R
\name{cdo_hurr}
\alias{cdo_hurr}
\title{hurr}
\usage{
cdo_hurr(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Hurricane days index per time period
}
