% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ifnotthen.R
\name{cdo_ifnotthen}
\alias{cdo_ifnotthen}
\title{ifnotthen}
\usage{
cdo_ifnotthen(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Conditional select one field
}
