% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ifnotthenc.R
\name{cdo_ifnotthenc}
\alias{cdo_ifnotthenc}
\title{ifnotthenc}
\usage{
cdo_ifnotthenc(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}
}
\description{
Conditional select a constant
}
