% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ifthen.R
\name{cdo_ifthen}
\alias{cdo_ifthen}
\title{ifthen}
\usage{
cdo_ifthen(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Conditional select one field
}
