% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ifthenc.R
\name{cdo_ifthenc}
\alias{cdo_ifthenc}
\title{ifthenc}
\usage{
cdo_ifthenc(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}
}
\description{
Conditional select a constant
}
