% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ifthenelse.R
\name{cdo_ifthenelse}
\alias{cdo_ifthenelse}
\title{ifthenelse}
\usage{
cdo_ifthenelse(ifile1, ifile2, ifile3, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2, ifile3}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Conditional select  two fields
}
