% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-import_amsr.R
\name{cdo_import_amsr}
\alias{cdo_import_amsr}
\title{import_amsr}
\usage{
cdo_import_amsr(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Import AMSR binary files
}
