% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-import_binary.R
\name{cdo_import_binary}
\alias{cdo_import_binary}
\title{import_binary}
\usage{
cdo_import_binary(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Import binary data sets
}
