% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-import_cmsaf.R
\name{cdo_import_cmsaf}
\alias{cdo_import_cmsaf}
\title{import_cmsaf}
\usage{
cdo_import_cmsaf(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Import CM
}
