% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-int.R
\name{cdo_int}
\alias{cdo_int}
\title{int}
\usage{
cdo_int(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
