% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-intlevel3d.R
\name{cdo_intlevel3d}
\alias{cdo_intlevel3d}
\title{intlevel3d}
\usage{
cdo_intlevel3d(ifile1, ifile2, tgtcoordinate = NULL, ofile = NULL)
}
\arguments{
\item{tgtcoordinate, }{STRING - filename for 3D vertical target coordinates}

\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Linear level interpolation from/to 3D vertical coordinates
}
