% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-intlevelx3d.R
\name{cdo_intlevelx3d}
\alias{cdo_intlevelx3d}
\title{intlevelx3d}
\usage{
cdo_intlevelx3d(ifile1, ifile2, tgtcoordinate = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{tgtcoordinate, }{STRING - filename for 3D vertical target coordinates}

\item{ofile}{String with the path to the output file.}
}
\description{
Linear level interpolation from/to 3D vertical coordinates
}
