% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-intntime.R
\name{cdo_intntime}
\alias{cdo_intntime}
\title{intntime}
\usage{
cdo_intntime(
  ifile,
  date = NULL,
  time = NULL,
  inc = NULL,
  n = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{date, }{STRING - Start date (format YYYY-MM-DD)}

\item{time, }{STRING - Start time (format hh:mm:ss)}

\item{inc, }{STRING - Optional increment (seconds, minutes, hours, days, months, years) [default: 0hour]}

\item{n, }{INTEGER - Number of timesteps from one timestep to the next}

\item{ofile}{String with the path to the output file.}
}
\description{
Time interpolation
}
