% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-intyear.R
\name{cdo_intyear}
\alias{cdo_intyear}
\title{intyear}
\usage{
cdo_intyear(ifile1, ifile2, years = NULL, obase = NULL)
}
\arguments{
\item{years, }{INTEGER - Comma-separated list or first/last\link{/inc} range of years}

\item{obase}{string with the basename of the output files.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}
}
\description{
Year interpolation
}
