% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-invertlev.R
\name{cdo_invertlev}
\alias{cdo_invertlev}
\title{invertlev}
\usage{
cdo_invertlev(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Invert levels
}
