% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-le.R
\name{cdo_le}
\alias{cdo_le}
\title{le}
\usage{
cdo_le(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Comparison of two fields
}
