% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-lec.R
\name{cdo_lec}
\alias{cdo_lec}
\title{lec}
\usage{
cdo_lec(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}
}
\description{
Comparison of a field with a constant
}
