% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-lt.R
\name{cdo_lt}
\alias{cdo_lt}
\title{lt}
\usage{
cdo_lt(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Comparison of two fields
}
