% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mastrfu.R
\name{cdo_mastrfu}
\alias{cdo_mastrfu}
\title{mastrfu}
\usage{
cdo_mastrfu(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mass stream function
}
