% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-max.R
\name{cdo_max}
\alias{cdo_max}
\title{max}
\usage{
cdo_max(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic on two datasets
}
