% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-maxc.R
\name{cdo_maxc}
\alias{cdo_maxc}
\title{maxc}
\usage{
cdo_maxc(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic with a constant
}
