% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mergegrid.R
\name{cdo_mergegrid}
\alias{cdo_mergegrid}
\title{mergegrid}
\usage{
cdo_mergegrid(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Merge grid
}
