% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mermax.R
\name{cdo_mermax}
\alias{cdo_mermax}
\title{mermax}
\usage{
cdo_mermax(ifile, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
Meridional statistics
}
