% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mermedian.R
\name{cdo_mermedian}
\alias{cdo_mermedian}
\title{mermedian}
\usage{
cdo_mermedian(ifile, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
Meridional statistics
}
