% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-merstd1.R
\name{cdo_merstd1}
\alias{cdo_merstd1}
\title{merstd1}
\usage{
cdo_merstd1(ifile, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{p, }{FLOAT - Percentile number in {0, ..., 100}}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Meridional statistics
}
