% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-monadd.R
\name{cdo_monadd}
\alias{cdo_monadd}
\title{monadd}
\usage{
cdo_monadd(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Monthly arithmetic
}
