% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mondiv.R
\name{cdo_mondiv}
\alias{cdo_mondiv}
\title{mondiv}
\usage{
cdo_mondiv(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Monthly arithmetic
}
