% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-monmax.R
\name{cdo_monmax}
\alias{cdo_monmax}
\title{monmax}
\usage{
cdo_monmax(ifile, complete_only = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{complete_only, }{BOOL - Process only complete years}

\item{ofile}{String with the path to the output file.}
}
\description{
Monthly statistics
}
