% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-monmul.R
\name{cdo_monmul}
\alias{cdo_monmul}
\title{monmul}
\usage{
cdo_monmul(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Monthly arithmetic
}
