% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-monpctl.R
\name{cdo_monpctl}
\alias{cdo_monpctl}
\title{monpctl}
\usage{
cdo_monpctl(ifile1, ifile2, ifile3, p = NULL, ofile = NULL)
}
\arguments{
\item{p, }{FLOAT - Percentile number in {0, ..., 100}}

\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Monthly percentile values
}
