% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-monrange.R
\name{cdo_monrange}
\alias{cdo_monrange}
\title{monrange}
\usage{
cdo_monrange(ifile, complete_only = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{complete_only, }{BOOL - Process only complete years}

\item{ofile}{String with the path to the output file.}
}
\description{
Monthly statistics
}
