% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mrotuvb.R
\name{cdo_mrotuvb}
\alias{cdo_mrotuvb}
\title{mrotuvb}
\usage{
cdo_mrotuvb(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Backward rotation of MPIOM data
}
