% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mul.R
\name{cdo_mul}
\alias{cdo_mul}
\title{mul}
\usage{
cdo_mul(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic on two datasets
}
