% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mulc.R
\name{cdo_mulc}
\alias{cdo_mulc}
\title{mulc}
\usage{
cdo_mulc(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Arithmetic with a constant
}
