% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-muldpy.R
\name{cdo_muldpy}
\alias{cdo_muldpy}
\title{muldpy}
\usage{
cdo_muldpy(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic with days
}
