% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-nint.R
\name{cdo_nint}
\alias{cdo_nint}
\title{nint}
\usage{
cdo_nint(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
