% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-not.R
\name{cdo_not}
\alias{cdo_not}
\title{not}
\usage{
cdo_not(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
