% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-pack.R
\name{cdo_pack}
\alias{cdo_pack}
\title{pack}
\usage{
cdo_pack(ifile, printparam = NULL, filename = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{printparam, }{BOOL - Print pack parameters to stdout for each variable}

\item{filename, }{STRING - Read pack parameters from file for each variable[format: name=<> add_offset=<> scale_factor=<>]}

\item{ofile}{String with the path to the output file.}
}
\description{
Pack data
}
