% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-pow.R
\name{cdo_pow}
\alias{cdo_pow}
\title{pow}
\usage{
cdo_pow(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
