% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-reci.R
\name{cdo_reci}
\alias{cdo_reci}
\title{reci}
\usage{
cdo_reci(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
