% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-reducegrid.R
\name{cdo_reducegrid}
\alias{cdo_reducegrid}
\title{reducegrid}
\usage{
cdo_reducegrid(ifile, mask = NULL, limitCoordsOutput = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{mask, }{STRING - file which holds the mask field}

\item{limitCoordsOutput, }{STRING - optional parameter to limit coordinates output: 'nobounds' disables coordinate bounds, 'nocoords' avoids all coordinate information}

\item{ofile}{String with the path to the output file.}
}
\description{
Reduce fields to user
}
