% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-regres.R
\name{cdo_regres}
\alias{cdo_regres}
\title{regres}
\usage{
cdo_regres(ifile, equal = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{equal, }{BOOL - Set to false for unequal distributed timesteps (default: true)}

\item{ofile}{String with the path to the output file.}
}
\description{
Regression
}
