% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remap.R
\name{cdo_remap}
\alias{cdo_remap}
\title{remap}
\usage{
cdo_remap(ifile, grid = NULL, weights = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Target grid description file or name}

\item{weights, }{STRING - Interpolation weights (SCRIP NetCDF file)}

\item{ofile}{String with the path to the output file.}
}
\description{
Grid remapping
}
