% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remapavg.R
\name{cdo_remapavg}
\alias{cdo_remapavg}
\title{remapavg}
\usage{
cdo_remapavg(ifile, grid = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Target grid description file or name}

\item{ofile}{String with the path to the output file.}
}
\description{
Remaps source points to target cells
}
