% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remapbic.R
\name{cdo_remapbic}
\alias{cdo_remapbic}
\title{remapbic}
\usage{
cdo_remapbic(ifile, grid = NULL, map3d = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{grid, }{STRING - Target grid description file or name}

\item{map3d, }{BOOL - Generate all mapfiles of the first 3D field}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Bicubic interpolation
}
