% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remapcon.R
\name{cdo_remapcon}
\alias{cdo_remapcon}
\title{remapcon}
\usage{
cdo_remapcon(ifile, grid = NULL, map3d = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Target grid description file or name}

\item{map3d, }{BOOL - Generate all mapfiles of the first 3D field}

\item{ofile}{String with the path to the output file.}
}
\description{
First order conservative remapping
}
