% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remapdis.R
\name{cdo_remapdis}
\alias{cdo_remapdis}
\title{remapdis}
\usage{
cdo_remapdis(ifile, grid = NULL, neighbors = NULL, map3d = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Target grid description file or name}

\item{neighbors, }{INTEGER - Number of nearest neighbors [default: 4]}

\item{map3d, }{BOOL - Generate all mapfiles of the first 3D field}

\item{ofile}{String with the path to the output file.}
}
\description{
Distance weighted average remapping
}
