% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remapeta.R
\name{cdo_remapeta}
\alias{cdo_remapeta}
\title{remapeta}
\usage{
cdo_remapeta(ifile, vct = NULL, oro = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{vct, }{STRING - File name of an ASCII dataset with the vertical coordinate table}

\item{oro, }{STRING - File name with the orography (surf. geopotential) of the target dataset (optional)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Remap vertical hybrid level
}
