% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remaplaf.R
\name{cdo_remaplaf}
\alias{cdo_remaplaf}
\title{remaplaf}
\usage{
cdo_remaplaf(ifile, grid = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Target grid description file or name}

\item{ofile}{String with the path to the output file.}
}
\description{
Largest area fraction remapping
}
