% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remapmean.R
\name{cdo_remapmean}
\alias{cdo_remapmean}
\title{remapmean}
\usage{
cdo_remapmean(ifile, grid = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{grid, }{STRING - Target grid description file or name}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Remaps source points to target cells
}
