% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remapmedian.R
\name{cdo_remapmedian}
\alias{cdo_remapmedian}
\title{remapmedian}
\usage{
cdo_remapmedian(ifile, grid = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Target grid description file or name}

\item{ofile}{String with the path to the output file.}
}
\description{
Remaps source points to target cells
}
