% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remaprange.R
\name{cdo_remaprange}
\alias{cdo_remaprange}
\title{remaprange}
\usage{
cdo_remaprange(ifile, grid = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{grid, }{STRING - Target grid description file or name}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Remaps source points to target cells
}
