% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-remapstd1.R
\name{cdo_remapstd1}
\alias{cdo_remapstd1}
\title{remapstd1}
\usage{
cdo_remapstd1(ifile, grid = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Target grid description file or name}

\item{ofile}{String with the path to the output file.}
}
\description{
Remaps source points to target cells
}
