% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-replace.R
\name{cdo_replace}
\alias{cdo_replace}
\title{replace}
\usage{
cdo_replace(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
The replace operator replaces variables of ifile1 with variables from ifile2 and write the result to ofile.
}
