% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-rhopot.R
\name{cdo_rhopot}
\alias{cdo_rhopot}
\title{rhopot}
\usage{
cdo_rhopot(ifile, pressure = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{pressure, }{FLOAT - Pressure in bar (constant value assigned to all levels)}

\item{ofile}{String with the path to the output file.}
}
\description{
Calculates potential density
}
