% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-runmean.R
\name{cdo_runmean}
\alias{cdo_runmean}
\title{runmean}
\usage{
cdo_runmean(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Running statistics
}
