% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-runmin.R
\name{cdo_runmin}
\alias{cdo_runmin}
\title{runmin}
\usage{
cdo_runmin(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Running statistics
}
