% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-runpctl.R
\name{cdo_runpctl}
\alias{cdo_runpctl}
\title{runpctl}
\usage{
cdo_runpctl(ifile, p = NULL, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Running percentile values
}
