% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-runvar1.R
\name{cdo_runvar1}
\alias{cdo_runvar1}
\title{runvar1}
\usage{
cdo_runvar1(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Running statistics
}
