% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seasmean.R
\name{cdo_seasmean}
\alias{cdo_seasmean}
\title{seasmean}
\usage{
cdo_seasmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Seasonal statistics
}
