% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seasmin.R
\name{cdo_seasmin}
\alias{cdo_seasmin}
\title{seasmin}
\usage{
cdo_seasmin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Seasonal statistics
}
