% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seaspctl.R
\name{cdo_seaspctl}
\alias{cdo_seaspctl}
\title{seaspctl}
\usage{
cdo_seaspctl(ifile1, ifile2, ifile3, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2, ifile3}{Strings with the path to the input files.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
Seasonal percentile values
}
