% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seasrange.R
\name{cdo_seasrange}
\alias{cdo_seasrange}
\title{seasrange}
\usage{
cdo_seasrange(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Seasonal statistics
}
