% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seasstd.R
\name{cdo_seasstd}
\alias{cdo_seasstd}
\title{seasstd}
\usage{
cdo_seasstd(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Seasonal statistics
}
