% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seasvar.R
\name{cdo_seasvar}
\alias{cdo_seasvar}
\title{seasvar}
\usage{
cdo_seasvar(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Seasonal statistics
}
