% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-selcircle.R
\name{cdo_selcircle}
\alias{cdo_selcircle}
\title{selcircle}
\usage{
cdo_selcircle(
  ifile,
  regions = NULL,
  lon = NULL,
  lat = NULL,
  radius = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{regions, }{STRING - Comma-separated list of ASCII formatted files with different regions}

\item{lon, }{FLOAT - Longitude of the center of the circle in degrees, default lon=0.0}

\item{lat, }{FLOAT - Latitude of the center of the circle in degrees, default lat=0.0}

\item{radius, }{STRING - Radius of the circle, default radius=1deg (units: deg, rad, km, m)}

\item{ofile}{String with the path to the output file.}
}
\description{
Select horizontal regions
}
