% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-select.R
\name{cdo_select}
\alias{cdo_select}
\title{select}
\usage{
cdo_select(
  ...,
  name = NULL,
  param = NULL,
  code = NULL,
  level = NULL,
  levrange = NULL,
  levidx = NULL,
  zaxisname = NULL,
  zaxisnum = NULL,
  ltype = NULL,
  gridname = NULL,
  gridnum = NULL,
  steptype = NULL,
  date = NULL,
  startdate = NULL,
  enddate = NULL,
  minute = NULL,
  hour = NULL,
  day = NULL,
  month = NULL,
  season = NULL,
  year = NULL,
  dom = NULL,
  timestep = NULL,
  timestep_of_year = NULL,
  timestepmask = NULL,
  ofile = NULL
)
}
\arguments{
\item{name, }{STRING - Comma-separated list of variable names.}

\item{param, }{STRING - Comma-separated list of parameter identifiers.}

\item{code, }{INTEGER - Comma-separated list or first/last\link{/inc} range of code numbers.}

\item{level, }{FLOAT - Comma-separated list of vertical levels.}

\item{levrange, }{FLOAT - First and last value of the level range.}

\item{levidx, }{INTEGER - Comma-separated list or first/last\link{/inc} range of index of levels.}

\item{zaxisname, }{STRING - Comma-separated list of zaxis names.}

\item{zaxisnum, }{INTEGER - Comma-separated list or first/last\link{/inc} range of zaxis numbers.}

\item{ltype, }{INTEGER - Comma-separated list or first/last\link{/inc} range of GRIB level types.}

\item{gridname, }{STRING - Comma-separated list of grid names.}

\item{gridnum, }{INTEGER - Comma-separated list or first/last\link{/inc} range of grid numbers.}

\item{steptype, }{STRING - Comma-separated list of timestep types (constant, avg, accum, min, max, range, diff, sum)}

\item{date, }{STRING - Comma-separated list of dates (format YYYY-MM-DDThh:mm:ss).}

\item{startdate, }{STRING - Start date (format YYYY-MM-DDThh:mm:ss).}

\item{enddate, }{STRING - End date (format YYYY-MM-DDThh:mm:ss).}

\item{minute, }{INTEGER - Comma-separated list or first/last\link{/inc} range of minutes.}

\item{hour, }{INTEGER - Comma-separated list or first/last\link{/inc} range of hours.}

\item{day, }{INTEGER - Comma-separated list or first/last\link{/inc} range of days.}

\item{month, }{INTEGER - Comma-separated list or first/last\link{/inc} range of months.}

\item{season, }{STRING - Comma-separated list of seasons (substring of DJFMAMJJASOND or ANN).}

\item{year, }{INTEGER - Comma-separated list or first/last\link{/inc} range of years.}

\item{dom, }{STRING - Comma-separated list of the day of month (e.g. 29feb).}

\item{timestep, }{INTEGER - Comma-separated list or first/last\link{/inc} range of timesteps. Negative values select timesteps from the end (NetCDF only).}

\item{timestep_of_year, }{INTEGER - Comma-separated list or first/last\link{/inc} range of timesteps of year.}

\item{timestepmask, }{STRING - Read timesteps from a mask file.}

\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Select fields
}
