% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sellonlatbox.R
\name{cdo_sellonlatbox}
\alias{cdo_sellonlatbox}
\title{sellonlatbox}
\usage{
cdo_sellonlatbox(
  ifile,
  lon1 = NULL,
  lon2 = NULL,
  lat1 = NULL,
  lat2 = NULL,
  idx1 = NULL,
  idx2 = NULL,
  idy1 = NULL,
  idy2 = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{lon1, }{FLOAT - Western longitude in degrees}

\item{lon2, }{FLOAT - Eastern longitude in degrees}

\item{lat1, }{FLOAT - Southern or northern latitude in degrees}

\item{lat2, }{FLOAT - Northern or southern latitude in degrees}

\item{idx1, }{INTEGER - Index of first longitude (1 - nlon)}

\item{idx2, }{INTEGER - Index of last longitude (1 - nlon)}

\item{idy1, }{INTEGER - Index of first latitude (1 - nlat)}

\item{idy2, }{INTEGER - Index of last latitude (1 - nlat)}

\item{ofile}{String with the path to the output file.}
}
\description{
Select a box
}
