% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-selmonth.R
\name{cdo_selmonth}
\alias{cdo_selmonth}
\title{selmonth}
\usage{
cdo_selmonth(
  ifile,
  timesteps = NULL,
  times = NULL,
  hours = NULL,
  days = NULL,
  months = NULL,
  years = NULL,
  seasons = NULL,
  startdate = NULL,
  enddate = NULL,
  nts1 = NULL,
  nts2 = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{timesteps, }{INTEGER - Comma-separated list or first/last\link{/inc} range of timesteps. Negative values select timesteps from the end (NetCDF only).}

\item{times, }{STRING - Comma-separated list of times (format hh:mm:ss).}

\item{hours, }{INTEGER - Comma-separated list or first/last\link{/inc} range of hours.}

\item{days, }{INTEGER - Comma-separated list or first/last\link{/inc} range of days.}

\item{months, }{INTEGER - Comma-separated list or first/last\link{/inc} range of months.}

\item{years, }{INTEGER - Comma-separated list or first/last\link{/inc} range of years.}

\item{seasons, }{STRING - Comma-separated list of seasons (substring of DJFMAMJJASOND or ANN).}

\item{startdate, }{STRING - Start date (format YYYY-MM-DDThh:mm:ss).}

\item{enddate, }{STRING - End date (format YYYY-MM-DDThh:mm:ss) \link{default: startdate}.}

\item{nts1, }{INTEGER - Number of timesteps before the selected month \link{default: 0}.}

\item{nts2, }{INTEGER - Number of timesteps after the selected month \link{default: nts1}.}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Select timesteps
}
