% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-selmulti.R
\name{cdo_selmulti}
\alias{cdo_selmulti}
\title{selmulti}
\usage{
cdo_selmulti(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Select multiple fields via GRIB1 parameters
}
