% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-selregion.R
\name{cdo_selregion}
\alias{cdo_selregion}
\title{selregion}
\usage{
cdo_selregion(
  ifile,
  regions = NULL,
  lon = NULL,
  lat = NULL,
  radius = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{regions, }{STRING - Comma-separated list of ASCII formatted files with different regions}

\item{lon, }{FLOAT - Longitude of the center of the circle in degrees, default lon=0.0}

\item{lat, }{FLOAT - Latitude of the center of the circle in degrees, default lat=0.0}

\item{radius, }{STRING - Radius of the circle, default radius=1deg (units: deg, rad, km, m)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Select horizontal regions
}
