% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seltimeidx.R
\name{cdo_seltimeidx}
\alias{cdo_seltimeidx}
\title{seltimeidx}
\usage{
cdo_seltimeidx(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Select timestep by index
}
