% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-selyearidx.R
\name{cdo_selyearidx}
\alias{cdo_selyearidx}
\title{selyearidx}
\usage{
cdo_selyearidx(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Select year by index
}
