% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setattribute.R
\name{cdo_setattribute}
\alias{cdo_setattribute}
\title{setattribute}
\usage{
cdo_setattribute(ifile, attributes = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{attributes, }{STRING - Comma-separated list of attributes.}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Set attributes
}
