% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setcindexbox.R
\name{cdo_setcindexbox}
\alias{cdo_setcindexbox}
\title{setcindexbox}
\usage{
cdo_setcindexbox(
  ifile,
  c = NULL,
  lon1 = NULL,
  lon2 = NULL,
  lat1 = NULL,
  lat2 = NULL,
  idx1 = NULL,
  idx2 = NULL,
  idy1 = NULL,
  idy2 = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c, }{FLOAT - Constant}

\item{lon1, }{FLOAT - Western longitude}

\item{lon2, }{FLOAT - Eastern longitude}

\item{lat1, }{FLOAT - Southern or northern latitude}

\item{lat2, }{FLOAT - Northern or southern latitude}

\item{idx1, }{INTEGER - Index of first longitude}

\item{idx2, }{INTEGER - Index of last longitude}

\item{idy1, }{INTEGER - Index of first latitude}

\item{idy2, }{INTEGER - Index of last latitude}

\item{ofile}{String with the path to the output file.}
}
\description{
Set a box to constant
}
