% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setcode.R
\name{cdo_setcode}
\alias{cdo_setcode}
\title{setcode}
\usage{
cdo_setcode(
  ifile,
  table = NULL,
  code = NULL,
  param = NULL,
  name = NULL,
  level = NULL,
  ltype = NULL,
  maxsteps = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{table, }{STRING - Parameter table file or name}

\item{code, }{INTEGER - Code number}

\item{param, }{STRING - Parameter identifier (GRIB1: code[.tabnum]; GRIB2: num[.cat[.dis]])}

\item{name, }{STRING - Variable name}

\item{level, }{FLOAT - New level}

\item{ltype, }{INTEGER - GRIB level type}

\item{maxsteps, }{INTEGER - Maximum number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Set field info
}
