% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setctomiss.R
\name{cdo_setctomiss}
\alias{cdo_setctomiss}
\title{setctomiss}
\usage{
cdo_setctomiss(
  ifile,
  neighbors = NULL,
  newmiss = NULL,
  c = NULL,
  rmin = NULL,
  rmax = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{neighbors, }{INTEGER - Number of nearest neighbors}

\item{newmiss, }{FLOAT - New missing value}

\item{c, }{FLOAT - Constant}

\item{rmin, }{FLOAT - Lower bound}

\item{rmax, }{FLOAT - Upper bound}

\item{ofile}{String with the path to the output file.}
}
\description{
Set missing value
}
