% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setday.R
\name{cdo_setday}
\alias{cdo_setday}
\title{setday}
\usage{
cdo_setday(
  ifile,
  day = NULL,
  month = NULL,
  year = NULL,
  units = NULL,
  date = NULL,
  time = NULL,
  inc = NULL,
  frequency = NULL,
  calendar = NULL,
  shiftValue = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{day, }{INTEGER - Value of the new day}

\item{month, }{INTEGER - Value of the new month}

\item{year, }{INTEGER - Value of the new year}

\item{units, }{STRING - Base units of the time axis (seconds, minutes, hours, days, months, years)}

\item{date, }{STRING - Date (format: YYYY-MM-DD)}

\item{time, }{STRING - Time (format: hh:mm:ss)}

\item{inc, }{STRING - Optional increment (seconds, minutes, hours, days, months, years) [default: 1hour]}

\item{frequency, }{STRING - Frequency of the time series (hour, day, month, year)}

\item{calendar, }{STRING - Calendar (standard, proleptic_gregorian, 360_day, 365_day, 366_day)}

\item{shiftValue, }{STRING - Shift value (e.g. -3hour)}

\item{ofile}{String with the path to the output file.}
}
\description{
Set time
}
