% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setgrid.R
\name{cdo_setgrid}
\alias{cdo_setgrid}
\title{setgrid}
\usage{
cdo_setgrid(
  ifile,
  grid = NULL,
  gridtype = NULL,
  gridarea = NULL,
  gridmask = NULL,
  projparams = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Grid description file or name}

\item{gridtype, }{STRING - Grid type (curvilinear, unstructured, regular, lonlat, projection or dereference)}

\item{gridarea, }{STRING - Data file, the first field is used as grid cell area}

\item{gridmask, }{STRING - Data file, the first field is used as grid mask}

\item{projparams, }{STRING - Proj library parameter (e.g.:+init=EPSG:3413)}

\item{ofile}{String with the path to the output file.}
}
\description{
Set grid information
}
