% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setgridarea.R
\name{cdo_setgridarea}
\alias{cdo_setgridarea}
\title{setgridarea}
\usage{
cdo_setgridarea(
  ifile,
  grid = NULL,
  gridtype = NULL,
  gridarea = NULL,
  gridmask = NULL,
  projparams = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{grid, }{STRING - Grid description file or name}

\item{gridtype, }{STRING - Grid type (curvilinear, unstructured, regular, lonlat, projection or dereference)}

\item{gridarea, }{STRING - Data file, the first field is used as grid cell area}

\item{gridmask, }{STRING - Data file, the first field is used as grid mask}

\item{projparams, }{STRING - Proj library parameter (e.g.:+init=EPSG:3413)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Set grid information
}
