% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setgridcell.R
\name{cdo_setgridcell}
\alias{cdo_setgridcell}
\title{setgridcell}
\usage{
cdo_setgridcell(ifile, value = NULL, cell = NULL, mask = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{value, }{FLOAT - Value of the grid cell}

\item{cell, }{INTEGER - Comma-separated list of grid cell indices}

\item{mask, }{STRING - Name of the data file which contains the mask}

\item{ofile}{String with the path to the output file.}
}
\description{
Set the value of a grid cell
}
