% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sethalo.R
\name{cdo_sethalo}
\alias{cdo_sethalo}
\title{sethalo}
\usage{
cdo_sethalo(
  ifile,
  east = NULL,
  west = NULL,
  south = NULL,
  north = NULL,
  value = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{east, }{INTEGER - East halo}

\item{west, }{INTEGER - West halo}

\item{south, }{INTEGER - South halo}

\item{north, }{INTEGER - North halo}

\item{value, }{FLOAT - Fill value (default is the missing value)}

\item{ofile}{String with the path to the output file.}
}
\description{
Set the bounds of a field
}
