% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setmisstodis.R
\name{cdo_setmisstodis}
\alias{cdo_setmisstodis}
\title{setmisstodis}
\usage{
cdo_setmisstodis(
  ifile,
  neighbors = NULL,
  newmiss = NULL,
  c = NULL,
  rmin = NULL,
  rmax = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{neighbors, }{INTEGER - Number of nearest neighbors}

\item{newmiss, }{FLOAT - New missing value}

\item{c, }{FLOAT - Constant}

\item{rmin, }{FLOAT - Lower bound}

\item{rmax, }{FLOAT - Upper bound}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Set missing value
}
