% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setpartabp.R
\name{cdo_setpartabp}
\alias{cdo_setpartabp}
\title{setpartabp}
\usage{
cdo_setpartabp(ifile, table = NULL, convert = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{table, }{STRING - Parameter table file or name}

\item{convert, }{STRING - Converts the units if necessary}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Set parameter table
}
