% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setrtoc2.R
\name{cdo_setrtoc2}
\alias{cdo_setrtoc2}
\title{setrtoc2}
\usage{
cdo_setrtoc2(
  ifile,
  oldval,
  newval,
  ... = NULL,
  rmin = NULL,
  rmax = NULL,
  c = NULL,
  c2 = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{oldval, newval, ..., }{FLOAT - Pairs of old and new values}

\item{rmin, }{FLOAT - Lower bound}

\item{rmax, }{FLOAT - Upper bound}

\item{c, }{FLOAT - New value - inside range}

\item{c2, }{FLOAT - New value - outside range}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Replace variable values
}
