% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-shiftx.R
\name{cdo_shiftx}
\alias{cdo_shiftx}
\title{shiftx}
\usage{
cdo_shiftx(ifile, nshift = NULL, cyclic = NULL, coord = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nshift, }{INTEGER - Number of grid cells to shift (default: 1)}

\item{cyclic, }{STRING - If set, cells are filled up cyclic (default: missing value)}

\item{coord, }{STRING - If set, coordinates are also shifted}

\item{ofile}{String with the path to the output file.}
}
\description{
Shift field
}
