% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-shifty.R
\name{cdo_shifty}
\alias{cdo_shifty}
\title{shifty}
\usage{
cdo_shifty(ifile, nshift = NULL, cyclic = NULL, coord = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{nshift, }{INTEGER - Number of grid cells to shift (default: 1)}

\item{cyclic, }{STRING - If set, cells are filled up cyclic (default: missing value)}

\item{coord, }{STRING - If set, coordinates are also shifted}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Shift field
}
