% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sin.R
\name{cdo_sin}
\alias{cdo_sin}
\title{sin}
\usage{
cdo_sin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
