% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sp2sp.R
\name{cdo_sp2sp}
\alias{cdo_sp2sp}
\title{sp2sp}
\usage{
cdo_sp2sp(ifile, trunc = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{trunc, }{INTEGER - New spectral resolution}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Spectral conversion
}
