% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-splitcode.R
\name{cdo_splitcode}
\alias{cdo_splitcode}
\title{splitcode}
\usage{
cdo_splitcode(ifile, swap = NULL, uuid = NULL, obase = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{swap, }{STRING - Swap the position of obase and xxx in the output filename}

\item{uuid, }{STRING - Add a UUID as global attribute <attname> to each output file}

\item{obase}{String with the basename of the output files.}
}
\description{
Split a dataset
}
