% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-splitlevel.R
\name{cdo_splitlevel}
\alias{cdo_splitlevel}
\title{splitlevel}
\usage{
cdo_splitlevel(ifile, swap = NULL, uuid = NULL, obase = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{swap, }{STRING - Swap the position of obase and xxx in the output filename}

\item{uuid, }{STRING - Add a UUID as global attribute <attname> to each output file}

\item{obase}{string with the basename of the output files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Split a dataset
}
