% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-splitsel.R
\name{cdo_splitsel}
\alias{cdo_splitsel}
\title{splitsel}
\usage{
cdo_splitsel(ifile, nsets = NULL, noffset = NULL, nskip = NULL, obase = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{nsets, }{INTEGER - Number of input timesteps for each output file}

\item{noffset, }{INTEGER - Number of input timesteps skipped before the first timestep range (optional)}

\item{nskip, }{INTEGER - Number of input timesteps skipped between timestep ranges (optional)}

\item{obase}{string with the basename of the output files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Split selected timesteps
}
