% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-splityear.R
\name{cdo_splityear}
\alias{cdo_splityear}
\title{splityear}
\usage{
cdo_splityear(ifile, format = NULL, obase = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{format, }{STRING - C-style format for strftime() (e.g. \%B for the full month name)}

\item{obase}{String with the basename of the output files.}
}
\description{
Split timesteps of a dataset
}
