% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sqr.R
\name{cdo_sqr}
\alias{cdo_sqr}
\title{sqr}
\usage{
cdo_sqr(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
