% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sqrt.R
\name{cdo_sqrt}
\alias{cdo_sqrt}
\title{sqrt}
\usage{
cdo_sqrt(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Mathematical functions
}
