% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-stdatm.R
\name{cdo_stdatm}
\alias{cdo_stdatm}
\title{stdatm}
\usage{
cdo_stdatm(
  ifile,
  const = NULL,
  seed = NULL,
  grid = NULL,
  start = NULL,
  end = NULL,
  inc = NULL,
  levels = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{const, }{FLOAT - Constant}

\item{seed, }{INTEGER - The seed for a new sequence of pseudo-random numbers \link{default: 1}}

\item{grid, }{STRING - Target grid description file or name}

\item{start, }{FLOAT - Start value of the loop}

\item{end, }{FLOAT - End value of the loop}

\item{inc, }{FLOAT - Increment of the loop \link{default: 1}}

\item{levels, }{FLOAT - Target levels in metre above surface}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Generate a field
}
