% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-strbre.R
\name{cdo_strbre}
\alias{cdo_strbre}
\title{strbre}
\usage{
cdo_strbre(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Strong breeze days index per time period
}
