% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-strgal.R
\name{cdo_strgal}
\alias{cdo_strgal}
\title{strgal}
\usage{
cdo_strgal(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Strong gale days index per time period
}
