% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-strwin.R
\name{cdo_strwin}
\alias{cdo_strwin}
\title{strwin}
\usage{
cdo_strwin(ifile, v = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{v, }{FLOAT - Horizontal wind speed threshold (m/s, default v = 10.5 m/s)}

\item{ofile}{String with the path to the output file.}
}
\description{
Strong wind days index per time period
}
