% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-subtrend.R
\name{cdo_subtrend}
\alias{cdo_subtrend}
\title{subtrend}
\usage{
cdo_subtrend(ifile1, ifile2, ifile3, equal = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2, ifile3}{Strings with the path to the input files.}

\item{equal, }{BOOL - Set to false for unequal distributed timesteps (default: true)}

\item{ofile}{String with the path to the output file.}
}
\description{
Add or subtract a trend
}
