% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-tee.R
\name{cdo_tee}
\alias{cdo_tee}
\title{tee}
\usage{
cdo_tee(ifile, outfile2 = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{outfile2, }{STRING - Destination filename for the copy of the input file}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Duplicate a data stream and write it to file
}
