% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timavg.R
\name{cdo_timavg}
\alias{cdo_timavg}
\title{timavg}
\usage{
cdo_timavg(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all timesteps
}
