% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timcor.R
\name{cdo_timcor}
\alias{cdo_timcor}
\title{timcor}
\usage{
cdo_timcor(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Correlation over time
}
