% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timcovar.R
\name{cdo_timcovar}
\alias{cdo_timcovar}
\title{timcovar}
\usage{
cdo_timcovar(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Covariance over time
}
