% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timcumsum.R
\name{cdo_timcumsum}
\alias{cdo_timcumsum}
\title{timcumsum}
\usage{
cdo_timcumsum(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Cumulative sum over all timesteps
}
