% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timmax.R
\name{cdo_timmax}
\alias{cdo_timmax}
\title{timmax}
\usage{
cdo_timmax(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all timesteps
}
