% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timmaxidx.R
\name{cdo_timmaxidx}
\alias{cdo_timmaxidx}
\title{timmaxidx}
\usage{
cdo_timmaxidx(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all timesteps
}
