% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timmin.R
\name{cdo_timmin}
\alias{cdo_timmin}
\title{timmin}
\usage{
cdo_timmin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all timesteps
}
