% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timminidx.R
\name{cdo_timminidx}
\alias{cdo_timminidx}
\title{timminidx}
\usage{
cdo_timminidx(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Statistical values over all timesteps
}
