% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timpctl.R
\name{cdo_timpctl}
\alias{cdo_timpctl}
\title{timpctl}
\usage{
cdo_timpctl(ifile1, ifile2, ifile3, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2, ifile3}{Strings with the path to the input files.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
Percentile values over all timesteps
}
