% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timrange.R
\name{cdo_timrange}
\alias{cdo_timrange}
\title{timrange}
\usage{
cdo_timrange(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all timesteps
}
