% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timselavg.R
\name{cdo_timselavg}
\alias{cdo_timselavg}
\title{timselavg}
\usage{
cdo_timselavg(ifile, nsets = NULL, noffset = NULL, nskip = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nsets, }{INTEGER - Number of input timesteps for each output timestep}

\item{noffset, }{INTEGER - Number of input timesteps skipped before the first timestep range (optional)}

\item{nskip, }{INTEGER - Number of input timesteps skipped between timestep ranges (optional)}

\item{ofile}{String with the path to the output file.}
}
\description{
Time range statistics
}
