% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timselmin.R
\name{cdo_timselmin}
\alias{cdo_timselmin}
\title{timselmin}
\usage{
cdo_timselmin(ifile, nsets = NULL, noffset = NULL, nskip = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{nsets, }{INTEGER - Number of input timesteps for each output timestep}

\item{noffset, }{INTEGER - Number of input timesteps skipped before the first timestep range (optional)}

\item{nskip, }{INTEGER - Number of input timesteps skipped between timestep ranges (optional)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Time range statistics
}
