% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timsort.R
\name{cdo_timsort}
\alias{cdo_timsort}
\title{timsort}
\usage{
cdo_timsort(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Timsort
}
