% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timstd.R
\name{cdo_timstd}
\alias{cdo_timstd}
\title{timstd}
\usage{
cdo_timstd(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all timesteps
}
