% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timvar.R
\name{cdo_timvar}
\alias{cdo_timvar}
\title{timvar}
\usage{
cdo_timvar(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all timesteps
}
