% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-topvalue.R
\name{cdo_topvalue}
\alias{cdo_topvalue}
\title{topvalue}
\usage{
cdo_topvalue(ifile, isovalue = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{isovalue, }{FLOAT - Isosurface value}

\item{ofile}{String with the path to the output file.}
}
\description{
Extract surface
}
