% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-install.R
\name{cdo_use}
\alias{cdo_use}
\title{Chose CDO version to use}
\usage{
cdo_use(version = c("system", "packaged"))
}
\arguments{
\item{version}{String with the cdo version to use:
\itemize{
\item \code{"system"} (the default) will use the system-wide installed version
(specifically, whatever path is returned by \code{Sys.which("cdo")}).
\item \code{"packaged"} instructs rcdo to use a package-specific version that can be
compiled and installed with \code{\link[=cdo_install]{cdo_install()}}.
}}
}
\value{
The path to the cdo executable (invisibly).
}
\description{
Chose CDO version to use
}
\details{
A one-time warning will be issued if the the cdo version found when
using \code{"system"} doesn't match the version used to build the rcdo package.
In that case, some operators documented in this package might not be
available to you or might behave slightly different.
However, most operators are stable, particularly the most often used ones.
}
