% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-uv2dv.R
\name{cdo_uv2dv}
\alias{cdo_uv2dv}
\title{uv2dv}
\usage{
cdo_uv2dv(ifile, gridtype = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{gridtype, }{STRING - Type of the grid: quadratic, linear, cubic (default: quadratic)}

\item{ofile}{String with the path to the output file.}
}
\description{
Wind transformation
}
