% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-uv2vr_cfd.R
\name{cdo_uv2vr_cfd}
\alias{cdo_uv2vr_cfd}
\title{uv2vr_cfd}
\usage{
cdo_uv2vr_cfd(
  ifile,
  u = NULL,
  v = NULL,
  boundOpt = NULL,
  outMode = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{u, }{STRING - Name of variable u (default: u)}

\item{v, }{STRING - Name of variable v (default: v)}

\item{boundOpt, }{INTEGER - Boundary condition option (0-3) (default: 0/1 for cyclic grids)}

\item{outMode, }{STRING - Output mode new/append (default: new)}

\item{ofile}{String with the path to the output file.}
}
\description{
Wind transformation
}
