% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varsavg.R
\name{cdo_varsavg}
\alias{cdo_varsavg}
\title{varsavg}
\usage{
cdo_varsavg(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Statistical values over all variables
}
