% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varsmax.R
\name{cdo_varsmax}
\alias{cdo_varsmax}
\title{varsmax}
\usage{
cdo_varsmax(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all variables
}
