% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varsmin.R
\name{cdo_varsmin}
\alias{cdo_varsmin}
\title{varsmin}
\usage{
cdo_varsmin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all variables
}
