% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varsstd.R
\name{cdo_varsstd}
\alias{cdo_varsstd}
\title{varsstd}
\usage{
cdo_varsstd(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all variables
}
