% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varsstd1.R
\name{cdo_varsstd1}
\alias{cdo_varsstd1}
\title{varsstd1}
\usage{
cdo_varsstd1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all variables
}
