% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varssum.R
\name{cdo_varssum}
\alias{cdo_varssum}
\title{varssum}
\usage{
cdo_varssum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all variables
}
