% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varsvar.R
\name{cdo_varsvar}
\alias{cdo_varsvar}
\title{varsvar}
\usage{
cdo_varsvar(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all variables
}
