% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varsvar1.R
\name{cdo_varsvar1}
\alias{cdo_varsvar1}
\title{varsvar1}
\usage{
cdo_varsvar1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all variables
}
