% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-vertmin.R
\name{cdo_vertmin}
\alias{cdo_vertmin}
\title{vertmin}
\usage{
cdo_vertmin(ifile, weights = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{weights, }{BOOL - weights=FALSE disables weighting by layer thickness [default: weights=TRUE]}

\item{ofile}{String with the path to the output file.}
}
\description{
Vertical statistics
}
