% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-vertrange.R
\name{cdo_vertrange}
\alias{cdo_vertrange}
\title{vertrange}
\usage{
cdo_vertrange(ifile, weights = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{weights, }{BOOL - weights=FALSE disables weighting by layer thickness [default: weights=TRUE]}

\item{ofile}{String with the path to the output file.}
}
\description{
Vertical statistics
}
