% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-vertstd.R
\name{cdo_vertstd}
\alias{cdo_vertstd}
\title{vertstd}
\usage{
cdo_vertstd(ifile, weights = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{weights, }{BOOL - weights=FALSE disables weighting by layer thickness \link{default: weights=TRUE}}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Vertical statistics
}
