% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydayavg.R
\name{cdo_ydayavg}
\alias{cdo_ydayavg}
\title{ydayavg}
\usage{
cdo_ydayavg(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
