% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydaymean.R
\name{cdo_ydaymean}
\alias{cdo_ydaymean}
\title{ydaymean}
\usage{
cdo_ydaymean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
