% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydaymin.R
\name{cdo_ydaymin}
\alias{cdo_ydaymin}
\title{ydaymin}
\usage{
cdo_ydaymin(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Multi
}
