% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydayrange.R
\name{cdo_ydayrange}
\alias{cdo_ydayrange}
\title{ydayrange}
\usage{
cdo_ydayrange(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
