% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydaystd.R
\name{cdo_ydaystd}
\alias{cdo_ydaystd}
\title{ydaystd}
\usage{
cdo_ydaystd(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
