% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydaystd1.R
\name{cdo_ydaystd1}
\alias{cdo_ydaystd1}
\title{ydaystd1}
\usage{
cdo_ydaystd1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
