% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydaysub.R
\name{cdo_ydaysub}
\alias{cdo_ydaysub}
\title{ydaysub}
\usage{
cdo_ydaysub(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
