% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydaysum.R
\name{cdo_ydaysum}
\alias{cdo_ydaysum}
\title{ydaysum}
\usage{
cdo_ydaysum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
