% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydayvar1.R
\name{cdo_ydayvar1}
\alias{cdo_ydayvar1}
\title{ydayvar1}
\usage{
cdo_ydayvar1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
