% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydrunmax.R
\name{cdo_ydrunmax}
\alias{cdo_ydrunmax}
\title{ydrunmax}
\usage{
cdo_ydrunmax(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
