% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydrunmin.R
\name{cdo_ydrunmin}
\alias{cdo_ydrunmin}
\title{ydrunmin}
\usage{
cdo_ydrunmin(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
