% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydrunpctl.R
\name{cdo_ydrunpctl}
\alias{cdo_ydrunpctl}
\title{ydrunpctl}
\usage{
cdo_ydrunpctl(ifile1, ifile2, ifile3, p = NULL, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2, ifile3}{Strings with the path to the input files.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
