% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydrunstd.R
\name{cdo_ydrunstd}
\alias{cdo_ydrunstd}
\title{ydrunstd}
\usage{
cdo_ydrunstd(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
