% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydrunsum.R
\name{cdo_ydrunsum}
\alias{cdo_ydrunsum}
\title{ydrunsum}
\usage{
cdo_ydrunsum(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
