% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydrunvar1.R
\name{cdo_ydrunvar1}
\alias{cdo_ydrunvar1}
\title{ydrunvar1}
\usage{
cdo_ydrunvar1(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
