% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yearadd.R
\name{cdo_yearadd}
\alias{cdo_yearadd}
\title{yearadd}
\usage{
cdo_yearadd(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Yearly arithmetic
}
