% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yearmonmean.R
\name{cdo_yearmonmean}
\alias{cdo_yearmonmean}
\title{yearmonmean}
\usage{
cdo_yearmonmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Yearly mean from monthly data
}
