% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yearmul.R
\name{cdo_yearmul}
\alias{cdo_yearmul}
\title{yearmul}
\usage{
cdo_yearmul(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Yearly arithmetic
}
