% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yearpctl.R
\name{cdo_yearpctl}
\alias{cdo_yearpctl}
\title{yearpctl}
\usage{
cdo_yearpctl(ifile1, ifile2, ifile3, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2, ifile3}{Strings with the path to the input files.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
Yearly percentile values
}
