% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yearsum.R
\name{cdo_yearsum}
\alias{cdo_yearsum}
\title{yearsum}
\usage{
cdo_yearsum(ifile, complete_only = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{complete_only, }{BOOL - Process only complete years}

\item{ofile}{String with the path to the output file.}
}
\description{
Yearly statistics
}
