% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhouradd.R
\name{cdo_yhouradd}
\alias{cdo_yhouradd}
\title{yhouradd}
\usage{
cdo_yhouradd(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
