% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhouravg.R
\name{cdo_yhouravg}
\alias{cdo_yhouravg}
\title{yhouravg}
\usage{
cdo_yhouravg(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
