% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhourmean.R
\name{cdo_yhourmean}
\alias{cdo_yhourmean}
\title{yhourmean}
\usage{
cdo_yhourmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
