% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhourmin.R
\name{cdo_yhourmin}
\alias{cdo_yhourmin}
\title{yhourmin}
\usage{
cdo_yhourmin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
