% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhourstd.R
\name{cdo_yhourstd}
\alias{cdo_yhourstd}
\title{yhourstd}
\usage{
cdo_yhourstd(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
