% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhourstd1.R
\name{cdo_yhourstd1}
\alias{cdo_yhourstd1}
\title{yhourstd1}
\usage{
cdo_yhourstd1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
