% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhoursub.R
\name{cdo_yhoursub}
\alias{cdo_yhoursub}
\title{yhoursub}
\usage{
cdo_yhoursub(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
