% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhoursum.R
\name{cdo_yhoursum}
\alias{cdo_yhoursum}
\title{yhoursum}
\usage{
cdo_yhoursum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
