% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhourvar.R
\name{cdo_yhourvar}
\alias{cdo_yhourvar}
\title{yhourvar}
\usage{
cdo_yhourvar(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
