% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhourvar1.R
\name{cdo_yhourvar1}
\alias{cdo_yhourvar1}
\title{yhourvar1}
\usage{
cdo_yhourvar1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
