% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ymonmax.R
\name{cdo_ymonmax}
\alias{cdo_ymonmax}
\title{ymonmax}
\usage{
cdo_ymonmax(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
