% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ymonmean.R
\name{cdo_ymonmean}
\alias{cdo_ymonmean}
\title{ymonmean}
\usage{
cdo_ymonmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
