% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ymonrange.R
\name{cdo_ymonrange}
\alias{cdo_ymonrange}
\title{ymonrange}
\usage{
cdo_ymonrange(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
