% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ymonstd.R
\name{cdo_ymonstd}
\alias{cdo_ymonstd}
\title{ymonstd}
\usage{
cdo_ymonstd(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
