% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ymonstd1.R
\name{cdo_ymonstd1}
\alias{cdo_ymonstd1}
\title{ymonstd1}
\usage{
cdo_ymonstd1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
