% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ymonsum.R
\name{cdo_ymonsum}
\alias{cdo_ymonsum}
\title{ymonsum}
\usage{
cdo_ymonsum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
