% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasmax.R
\name{cdo_yseasmax}
\alias{cdo_yseasmax}
\title{yseasmax}
\usage{
cdo_yseasmax(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
