% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasmean.R
\name{cdo_yseasmean}
\alias{cdo_yseasmean}
\title{yseasmean}
\usage{
cdo_yseasmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
