% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasmin.R
\name{cdo_yseasmin}
\alias{cdo_yseasmin}
\title{yseasmin}
\usage{
cdo_yseasmin(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Multi
}
