% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseaspctl.R
\name{cdo_yseaspctl}
\alias{cdo_yseaspctl}
\title{yseaspctl}
\usage{
cdo_yseaspctl(ifile, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{p, }{FLOAT - Percentile number in {0, ..., 100}}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Multi
}
