% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasrange.R
\name{cdo_yseasrange}
\alias{cdo_yseasrange}
\title{yseasrange}
\usage{
cdo_yseasrange(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
