% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasstd.R
\name{cdo_yseasstd}
\alias{cdo_yseasstd}
\title{yseasstd}
\usage{
cdo_yseasstd(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
