% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasstd1.R
\name{cdo_yseasstd1}
\alias{cdo_yseasstd1}
\title{yseasstd1}
\usage{
cdo_yseasstd1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
