% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseassub.R
\name{cdo_yseassub}
\alias{cdo_yseassub}
\title{yseassub}
\usage{
cdo_yseassub(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
