% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseassum.R
\name{cdo_yseassum}
\alias{cdo_yseassum}
\title{yseassum}
\usage{
cdo_yseassum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
