% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasvar.R
\name{cdo_yseasvar}
\alias{cdo_yseasvar}
\title{yseasvar}
\usage{
cdo_yseasvar(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
