% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasvar1.R
\name{cdo_yseasvar1}
\alias{cdo_yseasvar1}
\title{yseasvar1}
\usage{
cdo_yseasvar1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
