% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-zonpctl.R
\name{cdo_zonpctl}
\alias{cdo_zonpctl}
\title{zonpctl}
\usage{
cdo_zonpctl(ifile, p = NULL, zonaldes = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{zonaldes, }{STRING - Description of the zonal latitude bins needed for data on an unstructured grid. A predefined zonal description is zonal_<DY>. DY is the increment of the latitudes in degrees.}

\item{ofile}{String with the path to the output file.}
}
\description{
Zonal statistics
}
