% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-zonvar1.R
\name{cdo_zonvar1}
\alias{cdo_zonvar1}
\title{zonvar1}
\usage{
cdo_zonvar1(ifile, p = NULL, zonaldes = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{p, }{FLOAT - Percentile number in {0, ..., 100}}

\item{zonaldes, }{STRING - Description of the zonal latitude bins needed for data on an unstructured grid. A predefined zonal description is zonal_<DY>. DY is the increment of the latitudes in degrees.}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Zonal statistics
}
