% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIP_mergeModelFiles.R
\name{CMIP_mergeModelFiles}
\alias{CMIP_mergeModelFiles}
\title{merge_modelFiles}
\usage{
CMIP_mergeModelFiles(d, outdir = "ChinaHW_CMIP6_raw_bilinear", ...)
}
\arguments{
\item{d}{obj returned by \link{CMIP5Files_info}}

\item{...}{other parameters to \link{cdo_combine}}
}
\description{
merge_modelFiles
}
