% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-cmor.R
\name{cmor}
\alias{cmor}
\title{Climate Model Output Rewriting to produce CMIPcompliant data}
\description{
The CDO operator cmor converts an infile into a CMIP-compliant format by using the CMOR library. Each output file contains a single output variable. The name of the output files are generated by CMOR according to a template based on the DRS (Data reference Syntax) of the project. CMOR checks and applies the information delivered through the project dependend MIPtable on the infile. Additional information which is required for the conversion can be configured via keyvalues as optional parameters.  By specifying a variable selector keyvalue, e.g. cmor_name=tas, the user can pre-select a subset of infile variables. If name or code is specified, a corresponding cmor_name which can also be found in the MIPtable is also required to map the infile variable to the CMOR-variable. For mapping more variables at the operator call, one can specify a mapping table via keyword mapping_table.  Global attributes must be collected in info files and can be specified via keyword info. All required and optional global attributes as well as information about table file formats are given in the 'cdo cmor manual'.  If questions remain, do not hesitate to ask and send an email to wachsmannATdkrz.de.
}
