% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ifnotthenc.R, R/cdo-ifthenc.R,
%   R/family-condc.R
\name{cdo_ifnotthenc}
\alias{cdo_ifnotthenc}
\alias{cdo_ifthenc}
\alias{condc}
\title{Conditional select a constant}
\usage{
cdo_ifnotthenc(ifile, c = NULL, ofile = NULL)

cdo_ifthenc(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c}{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}
}
\description{
This module creates fields with a constant value or missing value. The fields in infile are handled as a mask. A value not equal to zero is treated as \"true\ zero is treated as \"false\".
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ifthenc     If then constant
                     / c      if i(t,x) NE 0  AND  i(t,x) NE miss
            o(t,x) =
                     \\\\ miss   if i(t,x) EQ 0  OR   i(t,x) EQ miss
ifnotthenc  If not then constant
                     / c      if i(t,x) EQ 0  AND  i(t,x) NE miss
            o(t,x) =
                     \\\\ miss   if i(t,x) NE 0  OR   i(t,x) EQ miss
}\if{html}{\out{</div>}}
}
