% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-cat.R, R/cdo-copy.R, R/family-copy.R
\name{cdo_cat}
\alias{cdo_cat}
\alias{cdo_copy}
\alias{copy}
\title{Copy datasets}
\usage{
cdo_cat(ifiles, ofile = NULL)

cdo_copy(ifiles, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
This module contains operators to copy, clone or concatenate datasets. infiles is an arbitrary number of input files. All input files need to have the same structure with the same variables on different timesteps.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{copy   Copy datasets
       Copies all input datasets to outfile. 
clone  Clone datasets
       Copies all input datasets to outfile. In contrast to the copy operator, clone tries
       not to change the input data. GRIB records are neither decoded nor decompressed.
cat    Concatenate datasets
       Concatenates all input datasets and appends the result to the end 
       of outfile. If outfile does not exist it will be created.
}\if{html}{\out{</div>}}
}
