% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDO.R
\name{create_maskregion}
\alias{create_maskregion}
\title{create mask region}
\usage{
create_maskregion(
  input,
  crs,
  buffer,
  tolerance,
  output = NULL,
  return = FALSE,
  ...
)
}
\arguments{
\item{input}{sf polygon}

\item{crs}{set CRS of input sf file (see st_crs)}

\item{buffer}{add buffer around input sf file (see st_buffer)}

\item{tolerance}{simplify polygon by setting tolerance  (see st_simplify)}

\item{output}{output}

\item{return}{return output? True or False}

\item{...}{passes on to other functions}
}
\value{
The path to the output file.
}
\description{
takes an input sf polygon and creates a text list suitable for CDO "maskregion"
}
