% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDO.R
\name{create_sellonlatbox}
\alias{create_sellonlatbox}
\title{create lonlat box}
\usage{
create_sellonlatbox(input)
}
\arguments{
\item{input}{sf polygon}
}
\value{
Bounding box xmin xmax ymin ymax
}
\description{
takes an input sf polygon and creates a bounding box in the format of CDO
}
