% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-deltat.R, R/op-cdo-deltat.R
\name{deltat}
\alias{deltat}
\alias{cdo_deltat}
\title{Difference between timesteps}
\usage{
cdo_deltat(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This operator computes the difference between each timestep.
}
