% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_csu.R, R/family-ecacsu.R
\name{cdo_eca_csu}
\alias{cdo_eca_csu}
\alias{ecacsu}
\title{Consecutive summer days index per time period}
\usage{
cdo_eca_csu(ifile, T = NULL, N = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{T}{FLOAT - Temperature threshold (unit: °C; default: T = 25°C)}

\item{N}{INTEGER - Minimum number of days exceeded (default: N = 5)}

\item{ofile}{String with the path to the output file.}
}
\description{
Let infile be a time series of the daily maximum temperature TX, then the largest number of consecutive days where TX > T is counted. The number T is an optional parameter with default T = 25°C. Note that TN have to be given in units of Kelvin, whereas T have to be given in degrees Celsius. A further output variable is the number of summer periods of more than N days. The date information of a timestep in outfile is the date of the last contributing timestep in infile.
}
