% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_etr.R, R/family-ecaetr.R
\name{cdo_eca_etr}
\alias{cdo_eca_etr}
\alias{ecaetr}
\title{Intraperiod extreme temperature range}
\usage{
cdo_eca_etr(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Let infile1 and infile2 be time series of thr maximum and minimum temperature TX and TN, respectively. Then the extreme temperature range is the difference of the maximum of TX and the minimum of TN. Note that TX and TN have to be given in the same units. The date information of a timestep in outfile is the date of the last contributing timesteps in infile1 and infile2.
}
