% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-ecahd.R, R/op-cdo-eca_hd.R
\name{ecahd}
\alias{ecahd}
\alias{cdo_eca_hd}
\title{Heating degree days per time period}
\usage{
cdo_eca_hd(ifile, T1 = NULL, T2 = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{T1}{FLOAT - Temperature limit (unit: 째C; default: T1 = 17째C)}

\item{T2}{FLOAT - Temperature limit (unit: 째C; default: T2 = T1)}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
Let infile be a time series of the daily mean temperature TG, then the heating degree days are defined as the sum of T1 - TG, where only values TG < T2 are considered. If T1 and T2 are omitted, a temperature of 17째C is used for both parameters. If only T1 is given, T2 is set to T1. Note that TG have to be given in units of kelvin, whereas T1 and T2 have to be given in degrees Celsius. The date information of a timestep in outfile is the date of the last contributing timestep in infile.
}
