% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_id.R, R/cdo-etccdi_id.R,
%   R/family-ecaid.R
\name{cdo_eca_id}
\alias{cdo_eca_id}
\alias{cdo_etccdi_id}
\alias{ecaid}
\title{Ice days index per time period}
\usage{
cdo_eca_id(ifile, freq = NULL, ofile = NULL)

cdo_etccdi_id(ifile, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{freq}{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
Let infile be a time series of the daily maximum temperature TX, then the number of days where TX < 0 °C is counted. Note that TX have to be given in units of Kelvin. Parameter is a comma-separated list of "key=values" pairs.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{eca_id     Ice days index per time period
           The operator counts over the entire time series.
           The date information of a timestep in outfile is the date of
           the last contributing timestep in infile.
etccdi_id  Ice days index per time period
           The default output frequency is yearly.
           The date information of a timestep in outfile is the mid of
           the frequency interval.
}\if{html}{\out{</div>}}
}
