% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_r75p.R, R/family-ecar75p.R
\name{cdo_eca_r75p}
\alias{cdo_eca_r75p}
\alias{ecar75p}
\title{Moderate wet days wrt 75th percentile of reference period}
\usage{
cdo_eca_r75p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Let infile1 be a time series RR of the daily precipitation amount at wet days (precipitation >= 1 mm) and infile2 be the 75th percentile RRn75 of the daily precipitation amount at wet days for any period used as reference. Then the percentage of wet days with RR > RRn75 is calculated. RRn75 is calculated as the 75th percentile of all wet days of a given climate reference period. Usually infile2 is generated by the operator ydaypctl,75. The date information of a timestep in outfile is the date of the last contributing timestep in infile1.
}
