% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-ecar90ptot.R, R/op-cdo-eca_r90ptot.R
\name{ecar90ptot}
\alias{ecar90ptot}
\alias{cdo_eca_r90ptot}
\title{Precipitation percent due to R90p days}
\usage{
cdo_eca_r90ptot(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operators that don't return filenames return a character vector with the
string output.
}
\description{
Let infile1 be a time series RR of the daily precipitation amount at wet days (precipitation >= 1 mm) and infile2 be the 90th percentile RRn90 of the daily precipitation amount at wet days for any period used as reference. Then the ratio of the precipitation sum at wet days with RR > RRn90 to the total precipitation sum is calculated. RRn90 is calculated as the 90th percentile of all wet days of a given climate reference period. Usually infile2 is generated by the operator ydaypctl,90. The date information of a timestep in outfile is the date of the last contributing timestep in infile1.
}
