% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_rr1.R, R/family-ecarr1.R
\name{cdo_eca_rr1}
\alias{cdo_eca_rr1}
\alias{ecarr1}
\title{Wet days index per time period}
\usage{
cdo_eca_rr1(ifile, R = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{R, }{FLOAT - Precipitation threshold (unit: mm; default: R = 1 mm)}

\item{ofile}{String with the path to the output file.}
}
\description{
Let infile be a time series of the daily precipitation amount RR in \link{mm} (or alternatively in \link{kg m-2}), then the number of days where RR is at least R is counted. R is an optional parameter with default R = 1 mm. The date information of a timestep in outfile is the date of the last contributing timestep in infile.
}
