% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_rx1day.R, R/cdo-etccdi_rx1day.R,
%   R/family-ecarx1day.R
\name{cdo_eca_rx1day}
\alias{cdo_eca_rx1day}
\alias{cdo_etccdi_rx1day}
\alias{ecarx1day}
\title{Highest one day precipitation amount per time period}
\usage{
cdo_eca_rx1day(ifile, freq = NULL, ofile = NULL)

cdo_etccdi_rx1day(ifile, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{freq}{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Let infile be a time series of the daily precipitation amount RR, then the maximum of RR is written to outfile. If the optional parameter mode is set to 'm' the maximum daily precipitation amounts are determined for each month. Parameter is a comma-separated list of \"key=values\" pairs.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{eca_rx1day     Highest one day precipitation amount per time period
               The operator counts over the entire time series.
               The date information of a timestep in outfile is the date of
               the last contributing timestep in infile.
etccdi_rx1day  Maximum 1-day Precipitation
               The default output frequency is yearly.
               The date information of a timestep in outfile is the mid of
               the frequency interval.
}\if{html}{\out{</div>}}
}
