% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-ecasdii.R, R/op-cdo-eca_sdii.R
\name{ecasdii}
\alias{ecasdii}
\alias{cdo_eca_sdii}
\title{Simple daily intensity index per time period}
\usage{
cdo_eca_sdii(ifile, R = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{R}{FLOAT - Precipitation threshold (unit: mm; default: R = 1 mm)}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operators that don't return filenames return a character vector with the
string output.
}
\description{
Let infile be a time series of the daily precipitation amount RR, then the mean precipitation amount at wet days (RR >= R) is written to outfile. R is an optional parameter with default R = 1 mm. The date information of a timestep in outfile is the date of the last contributing timestep in infile.
}
