% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-ecatg90p.R, R/op-cdo-eca_tg90p.R
\name{ecatg90p}
\alias{ecatg90p}
\alias{cdo_eca_tg90p}
\title{Warm days percent wrt 90th percentile of reference period}
\usage{
cdo_eca_tg90p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operators that don't return filenames return a character vector with the
string output.
}
\description{
Let infile1 be a time series of the daily mean temperature TG, and infile2 be the 90th percentile TGn90 of daily mean temperatures for any period used as reference. Then the percentage of time where TG > TGn90 is calculated. TGn90 is calculated as the 90th percentile of daily mean temperatures of a five day window centred on each calendar day of a given climate reference period. Note that both TG and TGn90 have to be given in the same units. The date information of a timestep in outfile is the date of the last contributing timestep in infile1.
}
