% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-ecatn10p.R, R/op-cdo-eca_tn10p.R
\name{ecatn10p}
\alias{ecatn10p}
\alias{cdo_eca_tn10p}
\title{Cold nights percent wrt 10th percentile of reference period}
\usage{
cdo_eca_tn10p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
Let infile1 be a time serie of the daily minimum temperature TN, and infile2 be the 10th percentile TNn10 of daily minimum temperatures for any period used as reference. Then the percentage of time where TN < TNn10 is calculated. TNn10 is calculated as the 10th percentile of daily minimum temperatures of a five day window centred on each calendar day of a given climate reference period. Note that both TN and TNn10 have to be given in the same units. The date information of a timestep in outfile is the date of the last contributing timestep in infile1.
}
