% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-filedes.R
\name{filedes}
\alias{filedes}
\title{Dataset description}
\description{
This module provides operators to print meta information about a dataset. The printed meta-data depends on the chosen operator.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{partab    Parameter table
          Prints all available meta information of the variables.
codetab   Parameter code table
          Prints a code table with a description of all variables.
          For each variable the operator prints one line listing the
          code, name, description and units.
griddes   Grid description
          Prints the description of all grids.
zaxisdes  Z-axis description
          Prints the description of all z-axes.
vct       Vertical coordinate table
          Prints the vertical coordinate table.
}\if{html}{\out{</div>}}
}
