% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fldcor.R, R/family-fldcor.R
\name{cdo_fldcor}
\alias{cdo_fldcor}
\alias{fldcor}
\title{Correlation in grid space}
\usage{
cdo_fldcor(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
The correlation coefficient is a quantity that gives the quality of a least squares fitting to the original data. This operator correlates all gridpoints of two fields for each timestep. With  S(t) = {x, i_1(t,x) != missval and i_2(t,x) != missval}  it is  o(t,1) = Cor{(i_1(t,x), i_2(t,x)), x_1 < x <= x_n}  where w(x) are the area weights obtained by the input streams. For every timestep t only those field elements x belong to the sample, which have i_1(t,x) != missval and i_2(t,x) != missval.
}
