% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fourier.R, R/family-fourier.R
\name{cdo_fourier}
\alias{cdo_fourier}
\alias{fourier}
\title{Fourier transformation}
\usage{
cdo_fourier(ifile, epsilon = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{epsilon}{INTEGER - -1: forward transformation; 1: backward transformation}

\item{ofile}{String with the path to the output file.}
}
\description{
The fourier operator performs the fourier transformation or the inverse fourier transformation of all input fields. If the number of timesteps is a power of 2 then the algorithm of the Fast Fourier Transformation (FFT) is used.   If the input stream infile consists only of complex fields, then the fields of outfile, computed by cdo -f ext fourier,1 -fourier,-1 infile outfile are the same than that of infile. For real input files see function retocomplex.
}
\section{Note}{

Complex numbers can only be stored in NetCDF4 and EXTRA format.
}

