% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-gradsdes.R
\name{gradsdes}
\alias{gradsdes}
\title{GrADS data descriptor file}
\description{
Creates a GrADS data descriptor file. Supported file formats are GRIB1, NetCDF, SERVICE, EXTRA and IEG. For GRIB1 files the GrADS map file is also generated. For SERVICE and EXTRA files the grid have to be specified with the CDO option '-g <grid>'. This module takes infile in order to create filenames for the descriptor (infile.ctl) and the map (infile.gmp) file.
}
