% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-intlevel3d.R, R/op-cdo-intlevel3d.R,
%   R/op-cdo-intlevelx3d.R
\name{intlevel3d}
\alias{intlevel3d}
\alias{cdo_intlevel3d}
\alias{cdo_intlevelx3d}
\title{Linear level interpolation from/to 3D vertical coordinates}
\usage{
cdo_intlevel3d(ifile1, ifile2, tgtcoordinate = NULL, ofile = NULL)

cdo_intlevelx3d(ifile1, ifile2, tgtcoordinate = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{tgtcoordinate}{STRING - filename for 3D vertical target coordinates}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This operator performs a linear vertical interpolation of 3D variables fields with given 3D vertical coordinates. infile1 contains the 3D data variables and infile2 the 3D vertical source coordinate. The parameter tgtcoordinate is a datafile with the 3D vertical target coordinate.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{intlevel3d   Linear level interpolation onto a 3D vertical coordinate
intlevelx3d  like intlevel3d but with extrapolation
}\if{html}{\out{</div>}}
}
