% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-intyear.R, R/family-intyear.R
\name{cdo_intyear}
\alias{cdo_intyear}
\alias{intyear}
\title{Year interpolation}
\usage{
cdo_intyear(ifile1, ifile2, years = NULL, obase = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{years}{INTEGER - Comma-separated list or first/last[/inc] range of years}

\item{obase}{String with the basename of the output files.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This operator performs linear interpolation between two years, timestep by timestep. The input files need to have the same structure with the same variables. The output files will be named <obase><yyyy><suffix> where yyyy will be the year and suffix is the filename extension derived from the file format.
}
\section{Note}{

This operator needs to open all output files simultaneously. The maximum number of open files depends on the operating system!
}

