% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-invertlev.R, R/family-invertlev.R
\name{cdo_invertlev}
\alias{cdo_invertlev}
\alias{invertlev}
\title{Invert levels}
\usage{
cdo_invertlev(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator inverts the levels of all 3D variables.
}
