% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-graph.R, R/family-maggraph.R
\name{cdo_graph}
\alias{cdo_graph}
\alias{maggraph}
\title{Line graph plot}
\usage{
cdo_graph(ifiles, parameter = NULL, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{parameter}{STRING - Comma-separated list of plot parameters}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator generates line graph plots. The data for the plot is read from infiles. The result is written to outfile. The default output file format is postscript, this can be changed with the device parameter.  Here is a list of all graph plot parameters:  Keyname    & Type    & Description device     & STRING  & Output device (ps, eps, pdf, png, gif, gif_animation, jpeg, svg, kml) ymin       & FLOAT   & Minimum value of the y-axis data ymax       & FLOAT   & Maximum value of the y-axis data linewidth  & INT     & Linewidth (default 8) stat       & STRING  & \"TRUE\" or \"FALSE\ to switch on the mean computation. Default is \"FALSE\". &         & Will be overridden to \"FALSE\ if input files have unequal number of time &         & steps or different start/end times. sigma      & FLOAT   & Standard deviation value for generating shaded back ground around the mean value. &         & To be used in conjunction with 'stat=\"TRUE\"' obsv       & STRING  & To indicate if the input files have an observation data, by setting to \"TRUE\". &         & Default value is \"FALSE\". The observation data should be the first file in the &         & input file list. The observation data is always plotted in black colour.
}
