% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIP6_process.R
\name{merge_modelFiles}
\alias{merge_modelFiles}
\title{merge_modelFiles}
\usage{
merge_modelFiles(d, outdir = "ChinaHW_CMIP6_raw_bilinear")
}
\arguments{
\item{d}{obj returned by \code{\link[=CMIP5Files_info]{CMIP5Files_info()}}}
}
\description{
merge_modelFiles
}
