% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatology.R
\name{nc_climatology}
\alias{nc_climatology}
\title{Calculates a climatology from a netcdf file}
\usage{
nc_climatology(ff, period = "seasonal", stat = "mean", vars = NULL,
  out_file = NULL, zip_file = FALSE, cdo_output = FALSE,
  overwrite = FALSE, ...)
}
\arguments{
\item{ff}{Target netcdf file}

\item{period}{Type of climatology. Must be one of yearly, monthly or seasonal.}

\item{stat}{Statistic you want calculated. Options are min, max, range, sum, mean, var, std. Default is mean. Only min, max and mean allowed for yearly currently.}

\item{vars}{Select the variables you want. If this is not given, stats will be calculated for all variables.}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{zip_file}{Do you want any output file to be zipped to save space? Default is FALSE.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}

\item{overwrite}{Do you want to overwrite out_file if it exists? Defaults to FALSE}

\item{...}{optional arguments to be sent to nc_remap if you need to remap the netcdf prior to calculating the stat}
}
\value{
data frame or netcdf file.
}
\description{
This function allows you to calculate climatological monthly statistics. Data must be monthly
}
