% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{nc_clip}
\alias{nc_clip}
\title{clip ncdf file}
\usage{
nc_clip(ff, vars = NULL, lon_range = c(-180, 180), lat_range = c(-90,
  90), vert_range = NULL, date_range = NULL, months = NULL,
  years = NULL, out_file = NULL, cdo_output = FALSE)
}
\arguments{
\item{ff}{This is the file to clip. This must be the full system path to the file.}

\item{vars}{Select the variables you want. If this is not given, all variables will be regridded.}

\item{lon_range}{longitude range. c(min_longituse, max_longitude).}

\item{lat_range}{latitude range. c(min_latitude, max_latitude).}

\item{vert_range}{This is the range of depths for vertical clipping, if wanted. c(min_depth, max_depth)}

\item{date_range}{This is the range of dates you want. c(date_min, date_max). "day/month/year" character string format.}

\item{months}{Months you want. c(month_1, month_2,...)}

\item{years}{Months you want. c(year_1, year_2,...)}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}
}
\description{
This function allows you to clip a netcdf file to a specific latlon box
}
