% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{nc_dates}
\alias{nc_dates}
\title{get dates available in netcdf file}
\usage{
nc_dates(ff)
}
\arguments{
\item{ff}{This is the file to check. Must be the full system path to the file.}
}
\description{
This function will tell you the dates in a netcdf file
}
