% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depths.R
\name{nc_depths}
\alias{nc_depths}
\title{nc_depths}
\usage{
nc_depths(ff)
}
\arguments{
\item{ff}{This is the file to check.}
}
\description{
This function will tell you the depths available in a netcdf file
}
\examples{
# Get the depths in the sample file
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
nc_depths(ff)

}
