% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf_detection.R
\name{nc_detect}
\alias{nc_detect}
\title{nc_detect}
\usage{
nc_detect(directory = getwd(), var = NULL, recursive = TRUE)
}
\arguments{
\item{ff}{This is the file to check. Must be the full system path to the file.}

\item{vars}{a character string of the variable to search for}
}
\description{
This function will list the netcdf files in a directory that contain a specified variable
}
