% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{nc_expr}
\alias{nc_expr}
\title{Rename variables in a netcdf file}
\usage{
nc_expr(ff, expr = NULL, out_file = NULL, zip_file = FALSE,
  cdo_output = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{ff}{This is the file to analyze.}

\item{expr}{A cdo expression.}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{zip_file}{Do you want any output file to be zipped to save space. Default is FALSE.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}

\item{overwrite}{Do you want to overwrite out_file if it exists? Defaults to FALSE for safety}

\item{...}{optional arguments to be sent to nc_remap if you need to remap the netcdf prior to calculating the stat}

\item{na_value}{This is a value in the raw netcdf file that needs to be treated as an na.}
}
\value{
data frame or netcdf file.
}
\description{
This function allows you to rename variables in a netcdf file. This function requires that the none  #' of the variables in the function contains arithmetic variables.
}
