% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{nc_find}
\alias{nc_find}
\title{search netcdf files in folder}
\usage{
nc_find(directory = getwd(), var = NULL, recursive = TRUE)
}
\arguments{
\item{directory}{This is the directory to check.}

\item{vars}{a character string of the variable to search for}
}
\value{
list of files
}
\description{
This function will list the netcdf files in a directory that contain a specified variable
}
