% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{nc_format}
\alias{nc_format}
\title{Find the format of a netcdf file.}
\usage{
nc_format(ff)
}
\arguments{
\item{ff}{This is the file to check.}
}
\description{
This function will tell you the format of a netcdf file
}
\examples{
# Get the format of the sample file
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
nc_format(ff)
}
