% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{nc_grid}
\alias{nc_grid}
\title{Extract the horizontal grid from a netcdf file}
\usage{
nc_grid(ff, cdo_output = FALSE)
}
\arguments{
\item{ff}{This is the file to analyze.}

\item{cdo_output}{Do you want to show the cdo output? Set to TRUE in case you want to troubleshoot errors.}
}
\description{
This function longitude/latitude grid cells in a netcdf file
}
\examples{

# Get the horizontal grid of the sample file
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
nc_grid(ff)
}
