% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_size.R
\name{nc_gridsize}
\alias{nc_gridsize}
\title{Calculate number of cells in netcdf grid}
\usage{
nc_gridsize(ff, cdo_output = FALSE, date_range = NULL)
}
\arguments{
\item{ff}{This is the file to analyze.}

\item{cdo_output}{Do you want to show the cdo output? Set to TRUE in case you want to troubleshoot errors.}
}
\description{
This function provides the number of cells in the horizontal netcdf grid
}
\examples{
# Get the grid size of the sample file
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
nc_gridsize(ff)
}
