% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{nc_merge}
\alias{nc_merge}
\title{merge netcdf files}
\usage{
nc_merge(ff_list, merge = "merge", expr = NULL, out_file = NULL,
  cdo_output = TRUE, overwrite = FALSE)
}
\arguments{
\item{ff_list}{This is a list of files to merge}

\item{merge}{This is the merge type. Use "merge" to combine files, "mergetime" to merge based on time.}

\item{expr}{This is an expression to apply to the merged files.}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}

\item{overwrite}{Do you want to overwrite out_file if it exists? Defaults to FALSE}
}
\value{
data frame or netcdf file.
}
\description{
This function allows you to remap a netcdf horizontally and vertically to a specific latlon box
}
