% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vertmean.R
\name{nc_merge_vertmean}
\alias{nc_merge_vertmean}
\title{merge and calculate the vertical means for a list of ncdf files}
\usage{
nc_merge_vertmean(ff_list, vars = NULL, coords = NULL, vert_scale,
  merge = "merge", expr = NULL, remap_point = "pre",
  na_value = NULL, out_file = NULL, cdo_output = TRUE,
  overwrite = FALSE)
}
\arguments{
\item{ff_list}{This is a character vector containing the files to merge}

\item{vars}{The variables to select from the files for merging..}

\item{coords}{A lonlat grid to regrid to. Longitude/Latitude columns. No horizontal mapping occurs if not supplied.}

\item{vert_scale}{This is the vertical scale you want. c(min_depth, max_depth, vertical_resolution).}

\item{merge}{This is the merge type. Use "merge" to combine files, "mergetime" to merge based on time.}

\item{expr}{This is a cdo expression to apply to the merged files.}

\item{remap_point}{This is the remap_point. Set to "pre" if you want to remap the files before merging, or "post" if you want to remap post-merging. The default is pre as this insures against horizontal grids being slightly different.}

\item{na_value}{This is a value in the raw netcdf file that needs to be treated as an na.#'}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}

\item{overwrite}{Do you want to overwrite out_file if it exists? Defaults to FALSE}
}
\value{
data frame or netcdf file.
}
\description{
This function allows you to remap a netcdf horizontally and vertically to a specific latlon box
}
