% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_list.R, R/read_list_ersem.R
\name{nc_read_list}
\alias{nc_read_list}
\title{Read a list of netcdf file to a nested data frame. Experimental}
\usage{
nc_read_list(ff_list, vars = NULL, cdo_output = FALSE,
  dim_check = 1.5e+08)

nc_read_list(ff_list, vars = NULL, cdo_output = FALSE,
  dim_check = 1.5e+08)
}
\arguments{
\item{ff_list}{Files to read. A character vector.}

\item{vars}{A list of variables you want to read in. Character vector. Everything is read in if this is empty.}

\item{cdo_output}{Do you want to show the cdo output? Set to TRUE in case you want to troubleshoot errors.}

\item{dim_check}{The number of data points in the final data frame that will ask to continue. Set to NULL if you don't want to check.}

\item{ff_list}{Files to read. A character vector.}

\item{vars}{A list of variables you want to read in. Character vector. Everything is read in if this is empty.}

\item{cdo_output}{Do you want to show the cdo output? Set to TRUE in case you want to troubleshoot errors.}

\item{dim_check}{The number of data points in the final data frame that will ask to continue. Set to NULL if you don't want to check.}
}
\description{
This reads a list of netcdf files to a data frame. This function is strict and currently only works when the files are of identifical format, and it will fail for netcdf files with multiple times. Experimental feature, likely to change.

This reads a list of netcdf files to a data frame. This function is strict and currently only works when the files are of identifical format, and it will fail for netcdf files with multiple times. Experimental feature, likely to change.
}
