% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{nc_remap}
\alias{nc_remap}
\title{remap a ncdf file}
\usage{
nc_remap(ff, vars = NULL, coords = NULL, vert_depths = NULL,
  out_file = NULL, cdo_output = FALSE, remapping = "bil",
  na_value = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{ff}{This is the file to regrid.}

\item{vars}{Select the variables you want to regrid. If this is not given, all variables will be regridded.}

\item{coords}{A 2 column matrix or data frame of the form (longitude, latitude) with coordinates for regridding. This can be regular or irregular. The function will calculate which it is.}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}

\item{remapping}{The type of remapping. bil = bilinear. nn = nearest neighbour. dis = distance weighted.}

\item{na_value}{This is a value in the raw netcdf file that needs to be treated as an na.#'}

\item{overwrite}{Do you want to overwrite out_file if it exists? Defaults to FALSE}

\item{...}{optional arguments to be sent to nc_clip if you need to clip prior to remapping.#'}
}
\value{
data frame or netcdf file.
}
\description{
This function allows you to remap a netcdf horizontally and vertically to any set of latlon coordinates.
}
\examples{

# Remapping NOAA world ocean atlas data to the region around the UK
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
# remapping to 1 degree resolution across all depth layers
uk_coords <- expand.grid(Longitude = seq(-20, 10, 1), Latitude = seq(48, 62, 1))
nc_remap(ff, vars = "t_an", coords = uk_coords)

# remapping to 1 degree resolution for 5, 50 and 100 metres in the region around the uk
nc_remap(ff, vars = "t_an", coords = uk_coords, vert_depths = c(5, 50, 100))
}
