% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_devel.R
\name{nc_remap2}
\alias{nc_remap2}
\title{remap a ncdf file (development version)}
\usage{
nc_remap2(ff, vars = NULL, coords = NULL, date_range = NULL,
  months = NULL, years = NULL, out_file = NULL, vert_depths = NULL,
  remapping = "bil", cdo_output = FALSE)
}
\arguments{
\item{ff}{This is the file to regrid.}

\item{vars}{Select the variables you want to regrid. If this is not given, all variables will be regridded.}

\item{coords}{A 2 column matrix or data frame of the form (longitude, latitude) with coordinates for regridding. This can be regular or irregular. The function will calculate which it is.}

\item{date_range}{This is the range of dates you want. c(date_min, date_max). "day/month/year" character string format. Ignored if not supplied}

\item{months}{Months you want. c(month_1, month_2,...). Ignored if not supplied.}

\item{years}{Years you want. c(year_1, year_2,...). Ignored if not supplied.}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{remapping}{The type of remapping. bil = bilinear. nn = nearest neighbour. dis = distance weighted.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}
}
\value{
data frame or netcdf file.
}
\description{
This function allows you to remap a netcdf horizontally and vertically to a specific latlon box
}
\examples{

# Remapping NOAA world ocean atlas data to the region around the UK
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
# remapping to 1 degree resolution across all depth layers
uk_coords <- expand.grid(Longitude = seq(-20, 10, 1), Latitude = seq(48, 62, 1))
nc_remap2(ff, vars = "t_an", coords = uk_coords)

# remapping to 1 degree resolution for 5, 50 and 100 metres in the region around the uk
nc_remap2(ff, vars = "t_an", coords = uk_coords, vert_depths = c(5, 50, 100))
}
