% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{nc_variables}
\alias{nc_variables}
\title{get variables available in netcdf file}
\usage{
nc_variables(ff)
}
\arguments{
\item{ff}{This is the file to move. This must be the full system path to the file.}
}
\description{
This function will tell you the variables in a netcdf file
}
