% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{nc_years}
\alias{nc_years}
\title{Find years available in a netcdf file}
\usage{
nc_years(ff)
}
\arguments{
\item{ff}{This is the file to read years from.}
}
\description{
Find years available in a netcdf file
}
\examples{

# Get the years in the sample file
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
nc_years(ff)
}
