% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{nc_zip}
\alias{nc_zip}
\title{Zip a netcdf file}
\usage{
nc_zip(ff, out_file = ff, overwrite = FALSE)
}
\arguments{
\item{ff}{This is the file to zip.}

\item{out_file}{The name of the file output. Defaults to ff}

\item{overwrite}{Do you want to overwrite out_file if it exists? Defaults to FALSE for safety}
}
\value{
data frame or netcdf file.
}
\description{
Zip a netcdf file
}
