% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-outputgmt.R
\name{outputgmt}
\alias{outputgmt}
\title{GMT output}
\description{
This module prints the first field of the input dataset to standard output. The output can be used to generate 2D Lon/Lat plots with GMT. The format of the output depends on the chosen operator.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{gmtxyz    GMT xyz format
          The operator exports the first field to the GMT xyz ASCII format.
          The output can be used to create contour plots with the GMT module pscontour.
gmtcells  GMT multiple segment format
          The operator exports the first field to the GMT multiple segment ASCII format.
          The output can be used to create shaded gridfill plots with the GMT module psxy.
}\if{html}{\out{</div>}}
}
