% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDO.R
\name{read_cdo}
\alias{read_cdo}
\title{Read CDO Input Files}
\usage{
read_cdo(...)
}
\arguments{
\item{...}{passes on to other functions}
}
\value{
A list containing the command structure for CDO operations.
}
\description{
Initializes the command structure with the input files.
}
\examples{
\dontrun{
read_cdo("/path/to/infile1.nc")

read_cdo(infile) \%>\%
operator("maskregion", myregion) \%>\%
  write_cdo(outfile)
}
}
