% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-replace.R, R/family-replace.R
\name{cdo_replace}
\alias{cdo_replace}
\alias{replace}
\title{Replace variables}
\usage{
cdo_replace(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator replaces variables in infile1 by variables from infile2 and write the result to outfile. Both input datasets need to have the same number of timesteps. All variable names may only occur once!
}
