% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-rhopot.R, R/family-rhopot.R
\name{cdo_rhopot}
\alias{cdo_rhopot}
\alias{rhopot}
\title{Calculates potential density}
\usage{
cdo_rhopot(ifile, pressure = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{pressure, }{FLOAT - Pressure in bar (constant value assigned to all levels)}

\item{ofile}{String with the path to the output file.}
}
\description{
This is a special operator for the post processing of the ocean and sea ice model MPIOM. It calculates the sea water potential density (name=rhopoto; code=18). Required input fields are sea water in-situ temperature (name=to; code=20) and sea water salinity (name=sao; code=5). Pressure is calculated from the level information or can be specified by the optional parameter.
}
