% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-rotuvb.R, R/op-cdo-rotuvb.R
\name{rotuvb}
\alias{rotuvb}
\alias{cdo_rotuvb}
\title{Rotation}
\usage{
cdo_rotuvb(ifile, u = NULL, v = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{u}{STRING - Pairs of zonal and meridional velocity components (use variable names or code numbers)}

\item{v}{STRING - Pairs of zonal and meridional velocity components (use variable names or code numbers)}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This is a special operator for datsets with wind components on a rotated grid, e.g. data from the regional model REMO. It performs a backward transformation of velocity components U and V from a rotated spherical system to a geographical system.
}
\section{Note}{

This is a specific implementation for data from the REMO model, it may not work with data from other sources.
}

