% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-samplegrid.R, R/op-cdo-samplegrid.R
\name{samplegrid}
\alias{samplegrid}
\alias{cdo_samplegrid}
\title{Resample grid}
\usage{
cdo_samplegrid(ifile, factor = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{factor}{INTEGER - Resample factor, typically 2, which will half the resolution}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This is a special operator for resampling the horizontal grid. No interpolation takes place. Resample factor=2 means every second grid point is removed. Only rectilinear and curvilinear source grids are supported by this operator.
}
