% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-delgridcell.R, R/cdo-selgridcell.R,
%   R/family-selgridcell.R
\name{cdo_delgridcell}
\alias{cdo_delgridcell}
\alias{cdo_selgridcell}
\alias{selgridcell}
\title{Select grid cells}
\usage{
cdo_delgridcell(ifile, indices = NULL, ofile = NULL)

cdo_selgridcell(ifile, indices = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{indices}{INTEGER - Comma-separated list or first/last[/inc] range of indices}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
The operator selects grid cells of all fields from infile. The user must specify the index of each grid cell. The resulting grid in outfile is unstructured.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{selgridcell  Select grid cells
delgridcell  Delete grid cells
}\if{html}{\out{</div>}}
}
