% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seltimeidx.R, R/family-seltimeidx.R
\name{cdo_seltimeidx}
\alias{cdo_seltimeidx}
\alias{seltimeidx}
\title{Select timestep by index}
\usage{
cdo_seltimeidx(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Selects field elements from infile2 according to a timestep index from infile1. The index of the timestep in infile1 should be the result of corresponding timminidx or timmaxidx operations, respectively.
}
