% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sethalo.R, R/family-sethalo.R
\name{cdo_sethalo}
\alias{cdo_sethalo}
\alias{sethalo}
\title{Set the bounds of a field}
\usage{
cdo_sethalo(
  ifile,
  east = NULL,
  west = NULL,
  south = NULL,
  north = NULL,
  value = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{east}{INTEGER - East halo}

\item{west}{INTEGER - West halo}

\item{south}{INTEGER - South halo}

\item{north}{INTEGER - North halo}

\item{value}{FLOAT - Fill value (default is the missing value)}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator sets the boundary in the east, west, south and north of the rectangular understood fields. Positive values of the parameters increase the boundary in the selected direction. Negative values decrease the field at the selected boundary. The new rows and columns are filled with the missing value. With the optional parameter value a different fill value can be used. Global cyclic fields are filled cyclically at the east and west borders, if the fill value is not set by the user. All input fields need to have the same horizontal grid.
}
