% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-shiftx.R, R/cdo-shifty.R,
%   R/family-shiftxy.R
\name{cdo_shiftx}
\alias{cdo_shiftx}
\alias{cdo_shifty}
\alias{shiftxy}
\title{Shift field}
\usage{
cdo_shiftx(ifile, nshift = NULL, cyclic = NULL, coord = NULL, ofile = NULL)

cdo_shifty(ifile, nshift = NULL, cyclic = NULL, coord = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nshift, }{INTEGER - Number of grid cells to shift (default: 1)}

\item{cyclic, }{STRING - If set, cells are filled up cyclic (default: missing value)}

\item{coord, }{STRING - If set, coordinates are also shifted}

\item{ofile}{String with the path to the output file.}
}
\description{
This module contains operators to shift all fields in x or y direction. All fields need to have the same horizontal rectilinear or curvilinear grid.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{shiftx  Shift x
        Shifts all fields in x direction.
shifty  Shift y
        Shifts all fields in y direction.
}\if{html}{\out{</div>}}
}
