% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-showattribute.R
\name{showattribute}
\alias{showattribute}
\title{Show attributes}
\description{
This operator prints the attributes of the data variables of a dataset.  Each attribute has the following structure:  [var_nm@][att_nm]  var_nm  Variable name (optional). Example: pressure att_nm  Attribute name (optional). Example: units  The value of var_nm is the name of the variable containing the attribute (named att_nm) that you want to print. Use wildcards to print the attribute att_nm of more than one variable. A value of var_nm of '\emph{' will print the attribute att_nm of all data variables. If var_nm is missing then att_nm refers to a global attribute.  The value of att_nm is the name of the attribute you want to print. Use wildcards to print more than one attribute. A value of att_nm of '}' will print all attributes.
}
