% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDO.R
\name{simplify_mask}
\alias{simplify_mask}
\title{simplify mask}
\usage{
simplify_mask(sf_in, crs = 4326, buffer = 8000, tolerance = 15000)
}
\arguments{
\item{sf_in}{sf polygon}

\item{crs}{set CRS of input sf file (see st_crs)}

\item{buffer}{add buffer around input sf file (see st_buffer)}

\item{tolerance}{simplify polygon by setting tolerance  (see st_simplify)}
}
\value{
output file.
}
\description{
simplify input sf for creating mask region
}
