% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-specconv.R, R/op-cdo-sp2sp.R
\name{specconv}
\alias{specconv}
\alias{cdo_sp2sp}
\title{Spectral conversion}
\usage{
cdo_sp2sp(ifile, trunc = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{trunc}{INTEGER - New spectral resolution}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
Changed the triangular truncation of all spectral fields. This operator performs downward conversion by cutting the resolution. Upward conversions are achieved by filling in zeros.
}
