% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-splitsel.R, R/family-splitsel.R
\name{cdo_splitsel}
\alias{cdo_splitsel}
\alias{splitsel}
\title{Split selected timesteps}
\usage{
cdo_splitsel(ifile, nsets = NULL, noffset = NULL, nskip = NULL, obase = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nsets, }{INTEGER - Number of input timesteps for each output file}

\item{noffset, }{INTEGER - Number of input timesteps skipped before the first timestep range (optional)}

\item{nskip, }{INTEGER - Number of input timesteps skipped between timestep ranges (optional)}

\item{obase}{String with the basename of the output files.}
}
\description{
This operator splits infile into pieces, one for each adjacent sequence t_1, ...., t_n of timesteps of the same selected time range. The output files will be named <obase><nnnnnn><suffix> where nnnnnn is the sequence number and suffix is the filename extension derived from the file format.
}
