% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timcovar.R, R/family-timcovar.R
\name{cdo_timcovar}
\alias{cdo_timcovar}
\alias{timcovar}
\title{Covariance over time}
\usage{
cdo_timcovar(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator calculates the covariance of two fields at each gridpoint over all timesteps. With  S(x) = \{t, i_1(t,x) != missval and i_2(t,x) != missval\}  it is  o(1,x) = Covar\{(i_1(t,x), i_2(t,x)), t_1 < t <= t_n\}  For every gridpoint x only those timesteps t belong to the sample, which have i_1(t,x) != missval and i_2(t,x) != missval.
}
