% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timcumsum.R, R/family-timcumsum.R
\name{cdo_timcumsum}
\alias{cdo_timcumsum}
\alias{timcumsum}
\title{Cumulative sum over all timesteps}
\usage{
cdo_timcumsum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
The timcumsum operator calculates the cumulative sum over all timesteps. Missing values are treated as numeric zero when summing.  o(t,x) = sum{i(t',x), 0<t'<=t}
}
