% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-timfillmiss.R
\name{timfillmiss}
\alias{timfillmiss}
\title{Temporal filling of missing values}
\description{
This operator fills in temporally missing values. The method parameter can be used to select the filling method. The default method=nearest fills missing values with the nearest neighbor value. Other options are forward and backward to fill missing values by forward or backward propagation of values. Use the limit parameter to set the maximum number of consecutive missing values to fill and max_gaps to set the maximum number of gaps to fill.
}
