% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timpctl.R, R/family-timpctl.R
\name{cdo_timpctl}
\alias{cdo_timpctl}
\alias{timpctl}
\title{Percentile values over all timesteps}
\usage{
cdo_timpctl(ifile1, ifile2, ifile3, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2, ifile3}{Strings with the path to the input files.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator computes percentiles over all timesteps in infile1. The algorithm uses histograms with minimum and maximum bounds given in infile2 and infile3, respectively. The default number of histogram bins is 101. The default can be overridden by defining the environment variable CDO_PCTL_NBINS. The files infile2 and infile3 should be the result of corresponding timmin and timmax operations, respectively. The time of outfile is determined by the time in the middle of all contributing timesteps of infile1. This can be change with the CDO option --timestat_date <first|middle|last>.  o(1,x) = pth percentile {i(t',x), t_1<t'<=t_n}
}
