% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timsort.R, R/family-timsort.R
\name{cdo_timsort}
\alias{cdo_timsort}
\alias{timsort}
\title{Timsort}
\usage{
cdo_timsort(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Sorts the elements in ascending order over all timesteps for every field position. After sorting it is:  o(t_1,x) <= o(t_2,x)      forall (t_1<t_2),x
}
