% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-unpack.R, R/family-unpack.R
\name{cdo_unpack}
\alias{cdo_unpack}
\alias{unpack}
\title{Unpack data}
\usage{
cdo_unpack(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Packing reduces the data volume by reducing the precision of the stored numbers. It is implemented using the NetCDF attributes add_offset and scale_factor. The operator unpack unpack all packed variables. The default data type for all variables is automatically changed to 32-bit floats. Use the CDO option -b F64 to change the data type to 64-bit floats, if needed.
}
