% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-verifygrid.R
\name{verifygrid}
\alias{verifygrid}
\title{Verify grid coordinates}
\description{
This operator verifies the coordinates of all horizontal grids found in infile. Among other things, it searches for duplicate cells, non-convex cells, and whether the center is located outside the cell bounds. Use the CDO option -v to output the position of these cells. This information can be useful to avoid problems when interpolating the data.
}
