% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dv2ps.R, R/family-wind2.R
\name{cdo_dv2ps}
\alias{cdo_dv2ps}
\alias{wind2}
\title{D and V to velocity potential and stream function}
\usage{
cdo_dv2ps(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Calculate spherical harmonic coefficients of velocity potential and stream function from spherical harmonic coefficients of relative divergence and vorticity. The divergence and vorticity need to have the names sd and svo or code numbers 155 and 138.
}
