% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDO.R
\name{write_cdo}
\alias{write_cdo}
\title{Write CDO Output File}
\usage{
write_cdo(cdo_command, outfile = NULL, debug = FALSE)
}
\arguments{
\item{cdo_command}{The command structure for CDO operations.}

\item{outfile}{The path for the output file (optional). If not specified, a temporary file will be created.}

\item{debug}{return just the CDO command, don't execute (TRUE/FALSE)}
}
\value{
The path to the output file.
}
\description{
Finalizes the command structure with the output file and executes the CDO command.
}
\examples{
\dontrun{
read_cdo("/path/to/infile.nc") \%>\%
  operator("maskregion", "/path/to/region.txt") \%>\%
  write_cdo("/path/to/outfile.nc")

read_cdo(infile) \%>\%
  operator("maskregion", myregion) \%>\%
  operator("sellonlatbox", lonlatbox) \%>\%
  write_cdo(outfile)
}
}
