% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yearadd.R, R/cdo-yeardiv.R,
%   R/cdo-yearmul.R, R/cdo-yearsub.R, R/family-yeararith.R
\name{cdo_yearadd}
\alias{cdo_yearadd}
\alias{cdo_yeardiv}
\alias{cdo_yearmul}
\alias{cdo_yearsub}
\alias{yeararith}
\title{Yearly arithmetic}
\usage{
cdo_yearadd(ifile1, ifile2, ofile = NULL)

cdo_yeardiv(ifile1, ifile2, ofile = NULL)

cdo_yearmul(ifile1, ifile2, ofile = NULL)

cdo_yearsub(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
This module performs simple arithmetic of a time series and one timestep with the same year. For each field in infile1 the corresponding field of the timestep in infile2 with the same year is used. The header information in infile1 have to be the same as in infile2. Usually infile2 is generated by an operator of the module YEARSTAT.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{yearadd  Add yearly time series
         Adds a time series and a yearly time series.
yearsub  Subtract yearly time series
         Subtracts a time series and a yearly time series.
yearmul  Multiply yearly time series
         Multiplies a time series and a yearly time series.
yeardiv  Divide yearly time series
         Divides a time series and a yearly time series.
}\if{html}{\out{</div>}}
}
