% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yearmonmean.R, R/family-yearmonstat.R
\name{cdo_yearmonmean}
\alias{cdo_yearmonmean}
\alias{yearmonstat}
\title{Yearly mean from monthly data}
\usage{
cdo_yearmonmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator computes the yearly mean of a monthly time series. Each month is weighted with the number of days per month. The time of outfile is determined by the time in the middle of all contributing timesteps of infile.  For every adjacent sequence t_1, ...,t_n of timesteps of the same year it is:  o(t,x) = mean\{i(t',x), t_1<t'<=t_n\}
}
