% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-ymonarith.R, R/op-cdo-ymonadd.R,
%   R/op-cdo-ymondiv.R, R/op-cdo-ymonmul.R, R/op-cdo-ymonsub.R
\name{ymonarith}
\alias{ymonarith}
\alias{cdo_ymonadd}
\alias{cdo_ymondiv}
\alias{cdo_ymonmul}
\alias{cdo_ymonsub}
\title{Multiyear monthly arithmetic}
\usage{
cdo_ymonadd(ifile1, ifile2, ofile = NULL)

cdo_ymondiv(ifile1, ifile2, ofile = NULL)

cdo_ymonmul(ifile1, ifile2, ofile = NULL)

cdo_ymonsub(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operators that don't return filenames return a character vector with the
string output.
}
\description{
This module performs simple arithmetic of a time series and one timestep with the same month of year. For each field in infile1 the corresponding field of the timestep in infile2 with the same month of year is used. The input files need to have the same structure with the same variables. Usually infile2 is generated by an operator of the module YMONSTAT.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ymonadd  Add multi-year monthly time series
         Adds a time series and a multi-year monthly time series.
ymonsub  Subtract multi-year monthly time series
         Subtracts a time series and a multi-year monthly time series.
ymonmul  Multiply multi-year monthly time series
         Multiplies a time series with a multi-year monthly time series.
ymondiv  Divide multi-year monthly time series
         Divides a time series by a multi-year monthly time series.
}\if{html}{\out{</div>}}
}
