% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-apply.R, R/family-apply.R
\name{cdo_apply}
\alias{cdo_apply}
\alias{apply}
\title{Apply operators}
\usage{
cdo_apply(ifile, operators = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{operators, }{STRING - Blank-separated list of CDO operators.}

\item{ofile}{String with the path to the output file.}
}
\description{
The apply utility runs the named operators on each input file. The input files must be enclosed in square brackets. This utility can only be used on a series of input files. These are all operators with more than one input file (infiles). Here is an incomplete list of these operators: copy, cat, merge, mergetime, select, ENSSTAT. The parameter operators is a blank-separated list of CDO operators. Use quotation marks if more than one operator is needed. Each operator may have only one input and output stream.
}
