% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-arithlat.R, R/op-cdo-divcoslat.R,
%   R/op-cdo-mulcoslat.R
\name{arithlat}
\alias{arithlat}
\alias{cdo_divcoslat}
\alias{cdo_mulcoslat}
\title{Arithmetic with latitude}
\usage{
cdo_divcoslat(ifile, ofile = NULL)

cdo_mulcoslat(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This module multiplies or divides each field element with the cosine of the latitude.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{mulcoslat  Multiply with the cosine of the latitude
           o(t,x) = i(t,x) * cos(latitude(x))
divcoslat  Divide by cosine of the latitude
           o(t,x) = i(t,x) / cos(latitude(x))
}\if{html}{\out{</div>}}
}
