% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-addc.R
\name{cdo_addc}
\alias{cdo_addc}
\title{addc}
\usage{
cdo_addc(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic with a constant
}
