% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-after.R
\name{cdo_after}
\alias{cdo_after}
\title{after}
\usage{
cdo_after(ifiles, vct = NULL, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{vct, }{STRING - File with VCT in ASCII format}

\item{ofile}{String with the path to the output file.}
}
\description{
ECHAM standard post processor
}
