% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-atan.R
\name{cdo_atan}
\alias{cdo_atan}
\title{atan}
\usage{
cdo_atan(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Mathematical functions
}
