% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-cat.R
\name{cdo_cat}
\alias{cdo_cat}
\title{cat}
\usage{
cdo_cat(ifiles, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Copy datasets
}
