% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-changemulti.R
\name{cdo_changemulti}
\alias{cdo_changemulti}
\title{changemulti}
\usage{
cdo_changemulti(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Select multiple fields via GRIB1 parameters
}
