% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-collgrid.R
\name{cdo_collgrid}
\alias{cdo_collgrid}
\title{collgrid}
\usage{
cdo_collgrid(ifiles, nx = NULL, names = NULL, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{nx, }{INTEGER - Number of regions in x direction [default: number of input files]}

\item{names, }{STRING - Comma-separated list of variable names [default: all variables]}

\item{ofile}{String with the path to the output file.}
}
\description{
Collect horizontal grid
}
