% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-utilize.R
\name{cdo_combine}
\alias{cdo_combine}
\alias{cdo_combine_large}
\title{cdo_combine}
\usage{
cdo_combine(
  files,
  outfile = NULL,
  outdir = ".",
  range = c(70, 140, 15, 55),
  delta = 3,
  cdo = "cdo",
  ncluster = 4,
  f_grid = "grid_d050.txt",
  verbose = TRUE,
  run = FALSE,
  ...
)

cdo_combine_large(files, outfile = NULL, chunk = 100, ...)
}
\description{
clip, mergetime and resample in one function
}
\note{
\if{html}{\out{<div class="sourceCode bash">}}\preformatted{export SKIP_SAME_TIME=1
}\if{html}{\out{</div>}}
}
\references{
\enumerate{
\item merge files with overlapping time periods, 2017,
https://code.mpimet.mpg.de/boards/1/topics/1134
}
}
