% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-consects.R
\name{cdo_consects}
\alias{cdo_consects}
\title{consects}
\usage{
cdo_consects(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Consecute timestep periods
}
