% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-delgridcell.R
\name{cdo_delgridcell}
\alias{cdo_delgridcell}
\title{delgridcell}
\usage{
cdo_delgridcell(ifile, indices = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{indices, }{INTEGER - Comma-separated list or first/last[/inc] range of indices}

\item{ofile}{String with the path to the output file.}
}
\description{
Select grid cells
}
