% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-delmulti.R
\name{cdo_delmulti}
\alias{cdo_delmulti}
\title{delmulti}
\usage{
cdo_delmulti(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Select multiple fields via GRIB1 parameters
}
