% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-deltat.R
\name{cdo_deltat}
\alias{cdo_deltat}
\title{deltat}
\usage{
cdo_deltat(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Difference between timesteps
}
