% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhouravg.R
\name{cdo_dhouravg}
\alias{cdo_dhouravg}
\title{dhouravg}
\usage{
cdo_dhouravg(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
