% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhourmin.R
\name{cdo_dhourmin}
\alias{cdo_dhourmin}
\title{dhourmin}
\usage{
cdo_dhourmin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
