% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dhourvar.R
\name{cdo_dhourvar}
\alias{cdo_dhourvar}
\title{dhourvar}
\usage{
cdo_dhourvar(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
