% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-distgrid.R
\name{cdo_distgrid}
\alias{cdo_distgrid}
\title{distgrid}
\usage{
cdo_distgrid(ifile, nx = NULL, ny = NULL, obase = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nx, }{INTEGER - Number of regions in x direction, or number of pieces for unstructured grids}

\item{ny, }{INTEGER - Number of regions in y direction [default: 1]}

\item{obase}{String with the basename of the output files.}
}
\description{
Distribute horizontal grid
}
