% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-divdpy.R
\name{cdo_divdpy}
\alias{cdo_divdpy}
\title{divdpy}
\usage{
cdo_divdpy(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic with days
}
