% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-duplicate.R
\name{cdo_duplicate}
\alias{cdo_duplicate}
\title{duplicate}
\usage{
cdo_duplicate(ifile, ndup = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ndup, }{INTEGER - Number of duplicates, default is 2.}

\item{ofile}{String with the path to the output file.}
}
\description{
Duplicates a dataset
}
