% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-dv2uv.R
\name{cdo_dv2uv}
\alias{cdo_dv2uv}
\title{dv2uv}
\usage{
cdo_dv2uv(ifile, gridtype = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{gridtype, }{STRING - Type of the grid: quadratic, linear, cubic (default: quadratic)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Wind transformation
}
