% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_cwd.R
\name{cdo_eca_cwd}
\alias{cdo_eca_cwd}
\title{eca_cwd}
\usage{
cdo_eca_cwd(ifile, R = NULL, N = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{R, }{FLOAT - Precipitation threshold (unit: mm; default: R = 1 mm)}

\item{N, }{INTEGER - Minimum number of days exceeded (default: N = 5)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Consecutive wet days index per time period
}
