% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_fd.R
\name{cdo_eca_fd}
\alias{cdo_eca_fd}
\title{eca_fd}
\usage{
cdo_eca_fd(ifile, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Frost days index per time period
}
