% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_gsl.R
\name{cdo_eca_gsl}
\alias{cdo_eca_gsl}
\title{eca_gsl}
\usage{
cdo_eca_gsl(ifile1, ifile2, nday = NULL, T = NULL, fland = NULL, ofile = NULL)
}
\arguments{
\item{nday, }{INTEGER - Number of consecutive days (default: nday = 6)}

\item{T, }{FLOAT - Temperature threshold (unit: °C; default: T = 5°C)}

\item{fland, }{FLOAT - Land fraction threshold (default: fland = 0.5)}

\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Thermal Growing season length index
}
