% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_pd.R
\name{cdo_eca_pd}
\alias{cdo_eca_pd}
\title{eca_pd}
\usage{
cdo_eca_pd(ifile, x = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{x, }{FLOAT - Daily precipitation amount threshold in [mm]}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Precipitation days index per time period
}
