% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_r75ptot.R
\name{cdo_eca_r75ptot}
\alias{cdo_eca_r75ptot}
\title{eca_r75ptot}
\usage{
cdo_eca_r75ptot(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Precipitation percent due to R75p days
}
