% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_r90p.R
\name{cdo_eca_r90p}
\alias{cdo_eca_r90p}
\title{eca_r90p}
\usage{
cdo_eca_r90p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Wet days wrt 90th percentile of reference period
}
