% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_r90ptot.R
\name{cdo_eca_r90ptot}
\alias{cdo_eca_r90ptot}
\title{eca_r90ptot}
\usage{
cdo_eca_r90ptot(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Precipitation percent due to R90p days
}
