% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_r95p.R
\name{cdo_eca_r95p}
\alias{cdo_eca_r95p}
\title{eca_r95p}
\usage{
cdo_eca_r95p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Very wet days wrt 95th percentile of reference period
}
