% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_tx10p.R
\name{cdo_eca_tx10p}
\alias{cdo_eca_tx10p}
\title{eca_tx10p}
\usage{
cdo_eca_tx10p(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Very cold days percent wrt 10th percentile of reference period
}
