% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensbrs.R
\name{cdo_ensbrs}
\alias{cdo_ensbrs}
\title{ensbrs}
\usage{
cdo_ensbrs(ifiles, obase = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{obase}{String with the basename of the output files.}
}
\description{
Ensemble validation tools
}
