% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-enscrps.R
\name{cdo_enscrps}
\alias{cdo_enscrps}
\title{enscrps}
\usage{
cdo_enscrps(ifiles, obase = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{obase}{String with the basename of the output files.}
}
\description{
Ensemble validation tools
}
