% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensrkhistspace.R
\name{cdo_ensrkhistspace}
\alias{cdo_ensrkhistspace}
\title{ensrkhistspace}
\usage{
cdo_ensrkhistspace(ifiles, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over an ensemble
}
