% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensrkhisttime.R
\name{cdo_ensrkhisttime}
\alias{cdo_ensrkhisttime}
\title{ensrkhisttime}
\usage{
cdo_ensrkhisttime(..., ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Statistical values over an ensemble
}
