% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ensstd1.R
\name{cdo_ensstd1}
\alias{cdo_ensstd1}
\title{ensstd1}
\usage{
cdo_ensstd1(ifiles, p = NULL, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over an ensemble
}
