% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eofcoeff.R
\name{cdo_eofcoeff}
\alias{cdo_eofcoeff}
\title{eofcoeff}
\usage{
cdo_eofcoeff(ifile1, ifile2, obase = NULL)
}
\arguments{
\item{obase}{string with the basename of the output files.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}
}
\description{
Principal coefficients of EOFs
}
