% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eoftime.R
\name{cdo_eoftime}
\alias{cdo_eoftime}
\title{eoftime}
\usage{
cdo_eoftime(ifile, neof = NULL, ofile1 = NULL, ofile2 = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{neof, }{INTEGER - Number of eigen functions}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Empirical Orthogonal Functions
}
