% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eqc.R
\name{cdo_eqc}
\alias{cdo_eqc}
\title{eqc}
\usage{
cdo_eqc(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}
}
\description{
Comparison of a field with a constant
}
