% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-etccdi_fd.R
\name{cdo_etccdi_fd}
\alias{cdo_etccdi_fd}
\title{etccdi_fd}
\usage{
cdo_etccdi_fd(ifile, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}
}
\description{
Frost days index per time period
}
