% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-etccdi_r1mm.R
\name{cdo_etccdi_r1mm}
\alias{cdo_etccdi_r1mm}
\title{etccdi_r1mm}
\usage{
cdo_etccdi_r1mm(ifile, x = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{x, }{FLOAT - Daily precipitation amount threshold in \link{mm}}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Precipitation days index per time period
}
