% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-etccdi_su.R
\name{cdo_etccdi_su}
\alias{cdo_etccdi_su}
\title{etccdi_su}
\usage{
cdo_etccdi_su(ifile, T = NULL, freq = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{T, }{FLOAT - Temperature threshold (unit: °C; default: T = 25°C)}

\item{freq, }{STRING - Output frequency (year, month)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Summer days index per time period
}
