% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.R
\name{cdo_execute}
\alias{cdo_execute}
\alias{cdo_execute_list}
\title{Execute a CDO operation}
\usage{
cdo_execute(
  operation,
  output = temp_output(operation),
  options = NULL,
  verbose = FALSE
)

cdo_execute_list(operations, output = NULL, options = NULL, verbose = FALSE)
}
\arguments{
\item{operation}{a CDO operation}

\item{output}{an output file or base string for output files. Defaults to
temporary files that will be deleted when its bond variable is garbage collected.}

\item{options}{character vector with CDO options.}

\item{verbose}{whether to print the command being executed.}

\item{operations}{a list of CDO operations}
}
\description{
Execute a CDO operation
}
