% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.R
\name{cdo_execute}
\alias{cdo_execute}
\title{Execute a CDO operation}
\usage{
cdo_execute(
  operation,
  output = temp_output(operation),
  options = NULL,
  verbose = FALSE
)
}
\arguments{
\item{operation}{a CDO operation}

\item{output}{an output file or base string for output files. Defaults to
temporary files.}

\item{options}{character vector with CDO options.}

\item{verbose}{whether to print the command being executed.}
}
\description{
Execute a CDO operation
}
