% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fdns.R
\name{cdo_fdns}
\alias{cdo_fdns}
\title{fdns}
\usage{
cdo_fdns(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Frost days where no snow index per time period
}
