% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fldcor.R
\name{cdo_fldcor}
\alias{cdo_fldcor}
\title{fldcor}
\usage{
cdo_fldcor(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Correlation in grid space
}
