% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-fldcovar.R
\name{cdo_fldcovar}
\alias{cdo_fldcovar}
\title{fldcovar}
\usage{
cdo_fldcovar(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Covariance in grid space
}
