% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gennn.R
\name{cdo_gennn}
\alias{cdo_gennn}
\title{gennn}
\usage{
cdo_gennn(ifile, grid = NULL, map3d = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{grid, }{STRING - Target grid description file or name}

\item{map3d, }{BOOL - Generate all mapfiles of the first 3D field}

\item{ofile}{String with the path to the output file.}
}
\description{
Nearest neighbor remapping
}
