% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-graph.R
\name{cdo_graph}
\alias{cdo_graph}
\title{graph}
\usage{
cdo_graph(ifiles, parameter = NULL, ofile = NULL)
}
\arguments{
\item{ifiles}{Character vector with the path to the input files.}

\item{parameter, }{STRING - Comma-separated list of plot parameters}

\item{ofile}{String with the path to the output file.}
}
\description{
Line graph plot
}
