% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-grfill.R
\name{cdo_grfill}
\alias{cdo_grfill}
\title{grfill}
\usage{
cdo_grfill(ifile, parameter = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{parameter, }{STRING - Comma-separated list of plot parameters}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Lat/Lon plot
}
