% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gridboxskew.R
\name{cdo_gridboxskew}
\alias{cdo_gridboxskew}
\title{gridboxskew}
\usage{
cdo_gridboxskew(ifile, nx = NULL, ny = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{nx, }{INTEGER - Number of grid boxes in x direction}

\item{ny, }{INTEGER - Number of grid boxes in y direction}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Statistical values over grid boxes
}
