% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gridboxvar1.R
\name{cdo_gridboxvar1}
\alias{cdo_gridboxvar1}
\title{gridboxvar1}
\usage{
cdo_gridboxvar1(ifile, nx = NULL, ny = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{nx, }{INTEGER - Number of grid boxes in x direction}

\item{ny, }{INTEGER - Number of grid boxes in y direction}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Statistical values over grid boxes
}
