% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-histsum.R
\name{cdo_histsum}
\alias{cdo_histsum}
\title{histsum}
\usage{
cdo_histsum(ifile, bounds = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{bounds, }{FLOAT - Comma-separated list of the bin bounds (-inf and inf valid)}

\item{ofile}{String with the path to the output file.}
}
\description{
Histogram
}
