% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ifnotthen.R
\name{cdo_ifnotthen}
\alias{cdo_ifnotthen}
\title{ifnotthen}
\usage{
cdo_ifnotthen(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Conditional select one field
}
