% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-input.R
\name{cdo_input}
\alias{cdo_input}
\title{input}
\usage{
cdo_input(ifile, grid = NULL, zaxis = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{Strings with the path to the input files.}

\item{grid, }{STRING - Grid description file or name}

\item{zaxis, }{STRING - Z-axis description file}

\item{ofile}{String with the path to the output file.}
}
\description{
Formatted input
}
