% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-install.R
\name{cdo_install}
\alias{cdo_install}
\title{Install the supported CDO version}
\usage{
cdo_install(
  reinstall = FALSE,
  proj = "/usr/",
  netcdf = "/usr/",
  fftw3 = "/usr/",
  eccodes = "/usr/"
)
}
\arguments{
\item{reinstall}{Logical. Set to true to force reinstallation.}

\item{proj, netcdf, fftw3, eccodes}{Location of the optional libraries.}
}
\description{
Install the supported CDO version
}
\details{
rcdo should work with your normal CDO installation but you if your installed
version is not the one used to generate this package, there could be some
small inconsistencies in the documentation, missing operators, extra operators
or changes in syntax.

\code{cdo_install()} will attempt to download, configure, compile and install
CDO version 2.4.3 in the package data directory. If this
version of CDO exists, the package will use it. Otherwise, it will use your
system's installation.
}
