% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-intlevel.R
\name{cdo_intlevel}
\alias{cdo_intlevel}
\title{intlevel}
\usage{
cdo_intlevel(
  ifile,
  level = NULL,
  zdescription = NULL,
  zvarname = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{level, }{FLOAT - Comma-separated list of target levels}

\item{zdescription, }{STRING - Path to a file containing a description of the Z-axis}

\item{zvarname, }{STRING - Use zvarname as the vertical 3D source coordinate instead of the 1D coordinate variable}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Linear level interpolation
}
