% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-invertlat.R
\name{cdo_invertlat}
\alias{cdo_invertlat}
\title{invertlat}
\usage{
cdo_invertlat(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Invert latitudes
}
