% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-isosurface.R
\name{cdo_isosurface}
\alias{cdo_isosurface}
\title{isosurface}
\usage{
cdo_isosurface(ifile, isovalue = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{isovalue, }{FLOAT - Isosurface value}

\item{ofile}{String with the path to the output file.}
}
\description{
Extract surface
}
