% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ln.R
\name{cdo_ln}
\alias{cdo_ln}
\title{ln}
\usage{
cdo_ln(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Mathematical functions
}
