% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-log10.R
\name{cdo_log10}
\alias{cdo_log10}
\title{log10}
\usage{
cdo_log10(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Mathematical functions
}
