% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ltc.R
\name{cdo_ltc}
\alias{cdo_ltc}
\title{ltc}
\usage{
cdo_ltc(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}
}
\description{
Comparison of a field with a constant
}
