% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-maskregion.R
\name{cdo_maskregion}
\alias{cdo_maskregion}
\title{maskregion}
\usage{
cdo_maskregion(ifile, regions = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{regions, }{STRING - Comma-separated list of ASCII formatted files with different regions}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Mask regions
}
