% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-merge.R
\name{cdo_merge}
\alias{cdo_merge}
\title{merge}
\usage{
cdo_merge(..., ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Merge files
}
