% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-merpctl.R
\name{cdo_merpctl}
\alias{cdo_merpctl}
\title{merpctl}
\usage{
cdo_merpctl(ifile, p = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{p, }{FLOAT - Percentile number in {0, ..., 100}}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Meridional statistics
}
