% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ml2pl.R
\name{cdo_ml2pl}
\alias{cdo_ml2pl}
\title{ml2pl}
\usage{
cdo_ml2pl(ifile, plevels = NULL, hlevels = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{plevels, }{FLOAT - Pressure levels in pascal}

\item{hlevels, }{FLOAT - Height levels in meter}

\item{ofile}{String with the path to the output file.}
}
\description{
Vertical interpolation
}
