% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mondiv.R
\name{cdo_mondiv}
\alias{cdo_mondiv}
\title{mondiv}
\usage{
cdo_mondiv(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Monthly arithmetic
}
