% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-muldpm.R
\name{cdo_muldpm}
\alias{cdo_muldpm}
\title{muldpm}
\usage{
cdo_muldpm(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Arithmetic with days
}
