% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-pack.R
\name{cdo_pack}
\alias{cdo_pack}
\title{pack}
\usage{
cdo_pack(ifile, printparam = NULL, filename = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{printparam, }{BOOL - Print pack parameters to stdout for each variable}

\item{filename, }{STRING - Read pack parameters from file for each variable\link{format: name=<> add_offset=<> scale_factor=<>}}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Pack data
}
