% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-replace.R
\name{cdo_replace}
\alias{cdo_replace}
\title{replace}
\usage{
cdo_replace(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Replace variables
}
