% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-rhopot.R
\name{cdo_rhopot}
\alias{cdo_rhopot}
\title{rhopot}
\usage{
cdo_rhopot(ifile, pressure = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{pressure, }{FLOAT - Pressure in bar (constant value assigned to all levels)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Calculates potential density
}
