% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-rotuvb.R
\name{cdo_rotuvb}
\alias{cdo_rotuvb}
\title{rotuvb}
\usage{
cdo_rotuvb(ifile, u, v, ... = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{u, v, ..., }{STRING - Pairs of zonal and meridional velocity components (use variable names or code numbers)}

\item{ofile}{String with the path to the output file.}
}
\description{
Rotation
}
