% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-runavg.R
\name{cdo_runavg}
\alias{cdo_runavg}
\title{runavg}
\usage{
cdo_runavg(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Running statistics
}
