% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-runsum.R
\name{cdo_runsum}
\alias{cdo_runsum}
\title{runsum}
\usage{
cdo_runsum(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Running statistics
}
