% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sealevelpressure.R
\name{cdo_sealevelpressure}
\alias{cdo_sealevelpressure}
\title{sealevelpressure}
\usage{
cdo_sealevelpressure(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Derived model parameters
}
