% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seasmax.R
\name{cdo_seasmax}
\alias{cdo_seasmax}
\title{seasmax}
\usage{
cdo_seasmax(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Seasonal statistics
}
