% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-seassum.R
\name{cdo_seassum}
\alias{cdo_seassum}
\title{seassum}
\usage{
cdo_seassum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Seasonal statistics
}
