% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.R
\name{cdo_set_output}
\alias{cdo_set_output}
\title{Set output and options}
\usage{
cdo_set_output(operation, output)
}
\arguments{
\item{operation}{a CDO operation}

\item{output}{an output file or base string for output files}
}
\description{
Set output and options
}
