% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sp2gp.R
\name{cdo_sp2gp}
\alias{cdo_sp2gp}
\title{sp2gp}
\usage{
cdo_sp2gp(ifile, type = NULL, trunc = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{type, }{STRING - Type of the grid: quadratic, linear, cubic (default: type=quadratic)}

\item{trunc, }{STRING - Triangular truncation}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Spectral transformation
}
