% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sp2sp.R
\name{cdo_sp2sp}
\alias{cdo_sp2sp}
\title{sp2sp}
\usage{
cdo_sp2sp(ifile, trunc = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{trunc, }{INTEGER - New spectral resolution}

\item{ofile}{String with the path to the output file.}
}
\description{
Spectral conversion
}
