% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-splitcode.R
\name{cdo_splitcode}
\alias{cdo_splitcode}
\title{splitcode}
\usage{
cdo_splitcode(ifile, swap = NULL, obase = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{swap, }{Swap the position of obase and xxx in the output filename}

\item{obase}{string with the basename of the output files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Split code numbers
}
