% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-splityearmon.R
\name{cdo_splityearmon}
\alias{cdo_splityearmon}
\title{splityearmon}
\usage{
cdo_splityearmon(ifile, format = NULL, obase = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{format, }{STRING - C-style format for strftime() (e.g. \%B for the full month name)}

\item{obase}{string with the basename of the output files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Split timesteps of a dataset
}
