% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-strbre.R
\name{cdo_strbre}
\alias{cdo_strbre}
\title{strbre}
\usage{
cdo_strbre(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Strong breeze days index per time period
}
