% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-subtrend.R
\name{cdo_subtrend}
\alias{cdo_subtrend}
\title{subtrend}
\usage{
cdo_subtrend(ifile, equal = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{equal, }{BOOL - Set to false for unequal distributed timesteps (default: true)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Add or subtract a trend
}
