% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timcumsum.R
\name{cdo_timcumsum}
\alias{cdo_timcumsum}
\title{timcumsum}
\usage{
cdo_timcumsum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Cumulative sum over all timesteps
}
