% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timmean.R
\name{cdo_timmean}
\alias{cdo_timmean}
\title{timmean}
\usage{
cdo_timmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Statistical values over all timesteps
}
