% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timminidx.R
\name{cdo_timminidx}
\alias{cdo_timminidx}
\title{timminidx}
\usage{
cdo_timminidx(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all timesteps
}
