% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timselpctl.R
\name{cdo_timselpctl}
\alias{cdo_timselpctl}
\title{timselpctl}
\usage{
cdo_timselpctl(
  ifile1,
  ifile2,
  ifile3,
  p = NULL,
  nsets = NULL,
  noffset = NULL,
  nskip = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile1, ifile2, ifile3}{Strings with the path to the input files.}

\item{p, }{FLOAT - Percentile number in \{0, ..., 100\}}

\item{nsets, }{INTEGER - Number of input timesteps for each output timestep}

\item{noffset, }{INTEGER - Number of input timesteps skipped before the first timestep range (optional)}

\item{nskip, }{INTEGER - Number of input timesteps skipped between timestep ranges (optional)}

\item{ofile}{String with the path to the output file.}
}
\description{
Time range percentile values
}
