% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timstd1.R
\name{cdo_timstd1}
\alias{cdo_timstd1}
\title{timstd1}
\usage{
cdo_timstd1(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all timesteps
}
