% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-trend.R
\name{cdo_trend}
\alias{cdo_trend}
\title{trend}
\usage{
cdo_trend(ifile, equal = NULL, ofile1 = NULL, ofile2 = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{equal, }{BOOL - Set to false for unequal distributed timesteps (default: true)}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Trend of time series
}
