% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varsavg.R
\name{cdo_varsavg}
\alias{cdo_varsavg}
\title{varsavg}
\usage{
cdo_varsavg(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all variables
}
