% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-varsmean.R
\name{cdo_varsmean}
\alias{cdo_varsmean}
\title{varsmean}
\usage{
cdo_varsmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Statistical values over all variables
}
