% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-vector.R
\name{cdo_vector}
\alias{cdo_vector}
\title{vector}
\usage{
cdo_vector(ifile, parameter = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{parameter, }{STRING - Comma-separated list of plot parameters}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Lat/Lon vector plot
}
