% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-wct.R
\name{cdo_wct}
\alias{cdo_wct}
\title{wct}
\usage{
cdo_wct(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Windchill temperature
}
