% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydaymax.R
\name{cdo_ydaymax}
\alias{cdo_ydaymax}
\title{ydaymax}
\usage{
cdo_ydaymax(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
