% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydaymean.R
\name{cdo_ydaymean}
\alias{cdo_ydaymean}
\title{ydaymean}
\usage{
cdo_ydaymean(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Multi
}
