% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydaymin.R
\name{cdo_ydaymin}
\alias{cdo_ydaymin}
\title{ydaymin}
\usage{
cdo_ydaymin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
