% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ydrunmean.R
\name{cdo_ydrunmean}
\alias{cdo_ydrunmean}
\title{ydrunmean}
\usage{
cdo_ydrunmean(ifile, nts = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{nts, }{INTEGER - Number of timesteps}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
