% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yeardiv.R
\name{cdo_yeardiv}
\alias{cdo_yeardiv}
\title{yeardiv}
\usage{
cdo_yeardiv(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
Yearly arithmetic
}
