% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yearmonmean.R
\name{cdo_yearmonmean}
\alias{cdo_yearmonmean}
\title{yearmonmean}
\usage{
cdo_yearmonmean(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Yearly mean from monthly data
}
