% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yearvar1.R
\name{cdo_yearvar1}
\alias{cdo_yearvar1}
\title{yearvar1}
\usage{
cdo_yearvar1(ifile, complete_only = NULL, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{complete_only, }{BOOL - Process only complete years}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Yearly statistics
}
