% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yhourmax.R
\name{cdo_yhourmax}
\alias{cdo_yhourmax}
\title{yhourmax}
\usage{
cdo_yhourmax(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
