% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ymonavg.R
\name{cdo_ymonavg}
\alias{cdo_ymonavg}
\title{ymonavg}
\usage{
cdo_ymonavg(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
