% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ymonmax.R
\name{cdo_ymonmax}
\alias{cdo_ymonmax}
\title{ymonmax}
\usage{
cdo_ymonmax(ifile, ofile = NULL)
}
\arguments{
\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{ofile}{String with the path to the output file.}

\item{obase}{string with the basename of the output files.}
}
\description{
Multi
}
