% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ymonmin.R
\name{cdo_ymonmin}
\alias{cdo_ymonmin}
\title{ymonmin}
\usage{
cdo_ymonmin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
