% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasmin.R
\name{cdo_yseasmin}
\alias{cdo_yseasmin}
\title{yseasmin}
\usage{
cdo_yseasmin(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
Multi
}
