% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-yseasmul.R
\name{cdo_yseasmul}
\alias{cdo_yseasmul}
\title{yseasmul}
\usage{
cdo_yseasmul(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ofile}{String with the path to the output file.}

\item{ifile, ifile1, ifile2}{String with the path to the input file(s).}

\item{obase}{string with the basename of the output files.}
}
\description{
Multi
}
