% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eqc.R, R/cdo-gec.R, R/cdo-gtc.R,
%   R/cdo-lec.R, R/cdo-ltc.R, R/cdo-nec.R, R/family-compc.R
\name{cdo_eqc}
\alias{cdo_eqc}
\alias{cdo_gec}
\alias{cdo_gtc}
\alias{cdo_lec}
\alias{cdo_ltc}
\alias{cdo_nec}
\alias{compc}
\title{Comparison of a field with a constant}
\usage{
cdo_eqc(ifile, c = NULL, ofile = NULL)

cdo_gec(ifile, c = NULL, ofile = NULL)

cdo_gtc(ifile, c = NULL, ofile = NULL)

cdo_lec(ifile, c = NULL, ofile = NULL)

cdo_ltc(ifile, c = NULL, ofile = NULL)

cdo_nec(ifile, c = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{c, }{FLOAT - Constant}

\item{ofile}{String with the path to the output file.}
}
\description{
This module compares all fields of a dataset with a constant. The resulting field is a mask containing 1 if the comparison is true and 0 if not. The type of comparison depends on the chosen operator.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{eqc  Equal constant
               /   1   if i(t,x) EQ c     AND  i(t,x),c NE miss
     o(t,x) = &lt;    0   if i(t,x) NE c     AND  i(t,x),c NE miss
               \\\\  miss if i(t,x) EQ miss  OR   c EQ miss
nec  Not equal constant
               /   1   if i(t,x) NE c     AND  i(t,x),c NE miss
     o(t,x) = &lt;    0   if i(t,x) EQ c     AND  i(t,x),c NE miss
               \\\\  miss if i(t,x) EQ miss  OR   c EQ miss
lec  Less equal constant
               /   1   if i(t,x) LE c     AND  i(t,x),c NE miss
     o(t,x) = &lt;    0   if i(t,x) GT c     AND  i(t,x),c NE miss
               \\\\  miss if i(t,x) EQ miss  OR   c EQ miss
ltc  Less than constant
               /   1   if i(t,x) LT c     AND  i(t,x),c NE miss
     o(t,x) = &lt;    0   if i(t,x) GE c     AND  i(t,x),c NE miss
               \\\\  miss if i(t,x) EQ miss  OR   c EQ miss
gec  Greater equal constant
               /   1   if i(t,x) GE c     AND  i(t,x),c NE miss
     o(t,x) = &lt;    0   if i(t,x) LT c     AND  i(t,x),c NE miss
               \\\\  miss if i(t,x) EQ miss  OR   c EQ miss
gtc  Greater than constant
               /   1   if i(t,x) GT c     AND  i(t,x),c NE miss
     o(t,x) = &lt;    0   if i(t,x) LE c     AND  i(t,x),c NE miss
               \\\\  miss if i(t,x) EQ miss  OR   c EQ miss
}\if{html}{\out{</div>}}
}
