% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-ifthenelse.R, R/family-cond2.R
\name{cdo_ifthenelse}
\alias{cdo_ifthenelse}
\alias{cond2}
\title{Conditional select  two fields}
\usage{
cdo_ifthenelse(ifile1, ifile2, ifile3, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2, ifile3}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator selects field elements from infile2 or infile3 with respect to infile1 and writes them to outfile. The fields in infile1 are handled as a mask. A value not equal to zero is treated as "true zero is treated as "false". The number of fields in infile1 has either to be the same as in infile2 or the same as in one timestep of infile2 or only one. infile2 and infile3 need to have the same number of fields. The fields in outfile inherit the meta data from infile2.  / i_2(t,x) if i_1(t,x) NE 0  AND  i_1(t,x) NE miss o(t,x) = <  i_3(t,x) if i_1(t,x) EQ 0  AND  i_1[t,x) NE miss miss     if i_1(t,x) EQ miss
}
