% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-eca_cfd.R, R/family-ecacfd.R
\name{cdo_eca_cfd}
\alias{cdo_eca_cfd}
\alias{ecacfd}
\title{Consecutive frost days index per time period}
\usage{
cdo_eca_cfd(ifile, N = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{N, }{INTEGER - Minimum number of days exceeded (default: N = 5)}

\item{ofile}{String with the path to the output file.}
}
\description{
Let infile be a time series of the daily minimum temperature TN, then the largest number of consecutive days where TN < 0 °C is counted. Note that TN have to be given in units of Kelvin. A further output variable is the number of frost periods of more than N days. The date information of a timestep in outfile is the date of the last contributing timestep in infile.
}
