% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-ecarr1.R, R/op-cdo-eca_rr1.R
\name{ecarr1}
\alias{ecarr1}
\alias{cdo_eca_rr1}
\title{Wet days index per time period}
\usage{
cdo_eca_rr1(ifile, R = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{R}{FLOAT - Precipitation threshold (unit: mm; default: R = 1 mm)}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
Let infile be a time series of the daily precipitation amount RR in [mm] (or alternatively in [kg m-2]), then the number of days where RR is at least R is counted. R is an optional parameter with default R = 1 mm. The date information of a timestep in outfile is the date of the last contributing timestep in infile.
}
