% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-filedes.R, R/op-cdo-codetab.R,
%   R/op-cdo-griddes.R, R/op-cdo-partab.R, R/op-cdo-vct.R, R/op-cdo-zaxisdes.R
\name{filedes}
\alias{filedes}
\alias{cdo_codetab}
\alias{cdo_griddes}
\alias{cdo_partab}
\alias{cdo_vct}
\alias{cdo_zaxisdes}
\title{Dataset description}
\usage{
cdo_codetab(ifile)

cdo_griddes(ifile)

cdo_partab(ifile)

cdo_vct(ifile)

cdo_zaxisdes(ifile)
}
\arguments{
\item{ifile}{String with the path to the input file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This module provides operators to print meta information about a dataset. The printed meta-data depends on the chosen operator.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{partab    Parameter table
          Prints all available meta information of the variables.
codetab   Parameter code table
          Prints a code table with a description of all variables.
          For each variable the operator prints one line listing the
          code, name, description and units.
griddes   Grid description
          Prints the description of all grids.
zaxisdes  Z-axis description
          Prints the description of all z-axes.
vct       Vertical coordinate table
          Prints the vertical coordinate table.
}\if{html}{\out{</div>}}
}
