% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-codetab.R, R/cdo-griddes.R, R/cdo-partab.R,
%   R/cdo-vct.R, R/cdo-zaxisdes.R, R/family-filedes.R
\name{cdo_codetab}
\alias{cdo_codetab}
\alias{cdo_griddes}
\alias{cdo_partab}
\alias{cdo_vct}
\alias{cdo_zaxisdes}
\alias{filedes}
\title{Dataset description}
\usage{
cdo_codetab(ifile)

cdo_griddes(ifile)

cdo_partab(ifile)

cdo_vct(ifile)

cdo_zaxisdes(ifile)
}
\arguments{
\item{ifile}{String with the path to the input file.}
}
\description{
This module provides operators to print meta information about a dataset. The printed meta-data depends on the chosen operator.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{partab    Parameter table
          Prints all available meta information of the variables.
codetab   Parameter code table
          Prints a code table with a description of all variables.
          For each variable the operator prints one line listing the
          code, name, description and units.
griddes   Grid description
          Prints the description of all grids.
zaxisdes  Z-axis description
          Prints the description of all z-axes.
vct       Vertical coordinate table
          Prints the vertical coordinate table.
}\if{html}{\out{</div>}}
}
