% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gridcellindex.R, R/family-getgridcell.R
\name{cdo_gridcellindex}
\alias{cdo_gridcellindex}
\alias{getgridcell}
\title{Get grid cell index}
\usage{
cdo_gridcellindex(ifile, lon = NULL, lat = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{lon}{INTEGER - Longitude of the grid cell in degree}

\item{lat}{INTEGER - Latitude of the grid cell in degree}
}
\description{
Get the grid cell index of one grid point selected by the parameter lon and lat.
}
