% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-healpix.R
\name{healpix}
\alias{healpix}
\title{Change healpix resolution}
\description{
Degrade or upgrade the resolution of a healpix grid.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{hpdegrade  Degrade healpix
           Degrade the resolution of a healpix grid. The value of the target pixel is the mean of the source pixels.
hpupgrade  Upgrade healpix
           Upgrade the resolution of a healpix grid. The values of the target pixels is the value of the source pixel.
}\if{html}{\out{</div>}}
}
