% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-invertlat.R, R/family-invert.R
\name{cdo_invertlat}
\alias{cdo_invertlat}
\alias{invert}
\title{Invert latitudes}
\usage{
cdo_invertlat(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator inverts the latitudes of all fields on a rectilinear grid.
}
