% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-invertlev.R, R/op-cdo-invertlev.R
\name{invertlev}
\alias{invertlev}
\alias{cdo_invertlev}
\title{Invert levels}
\usage{
cdo_invertlev(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This operator inverts the levels of all 3D variables.
}
