% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-mastrfu.R, R/op-cdo-mastrfu.R
\name{mastrfu}
\alias{mastrfu}
\alias{cdo_mastrfu}
\title{Mass stream function}
\usage{
cdo_mastrfu(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operators that don't return filenames return a character vector with the
string output.
}
\description{
This is a special operator for the post processing of the atmospheric general circulation model ECHAM. It computes the mass stream function (code=272). The input dataset have to be a zonal mean of v-velocity [m/s] (code=132) on pressure levels.
}
