% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-mastrfu.R, R/family-mastrfu.R
\name{cdo_mastrfu}
\alias{cdo_mastrfu}
\alias{mastrfu}
\title{Mass stream function}
\usage{
cdo_mastrfu(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\description{
This is a special operator for the post processing of the atmospheric general circulation model ECHAM. It computes the mass stream function (code=272). The input dataset have to be a zonal mean of v-velocity \link{m/s} (code=132) on pressure levels.
}
