% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{nc_cellareas}
\alias{nc_cellareas}
\title{grid cell areas of a ncdf file}
\usage{
nc_cellareas(ff, cdo_output = FALSE)
}
\arguments{
\item{ff}{This is the file to analyze.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}
}
\value{
data frame with the cell areas in square metres
}
\description{
Calculates the areas of each grid cell of a netcdf file.
}
\examples{

# Calculating cell areas for NOAA world ocean atlas data sample file
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
nc_cellareas(ff)
}
