% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{nc_clip}
\alias{nc_clip}
\title{Clip a netcdf file}
\usage{
nc_clip(ff, vars = NULL, lon_range = c(-180, 180), lat_range = c(-90,
  90), vert_range = NULL, date_range = NULL, months = NULL,
  years = NULL, out_file = NULL, cdo_output = FALSE,
  zip_file = FALSE, overwrite = FALSE)
}
\arguments{
\item{ff}{This is the file to clip.}

\item{vars}{Select the variables you want. If this is not given, all variables will be regridded.}

\item{lon_range}{longitude range. c(min_longituse, max_longitude).}

\item{lat_range}{latitude range. c(min_latitude, max_latitude).}

\item{vert_range}{This is the range of depths for vertical clipping, if wanted. c(min_depth, max_depth)}

\item{date_range}{This is the range of dates you want. c(date_min, date_max). "day/month/year" character string format.}

\item{months}{Months you want. c(month_1, month_2,...)}

\item{years}{Months you want. c(year_1, year_2,...)}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}

\item{zip_file}{Do you want any output file to be zipped to save space. Default is FALSE.}

\item{overwrite}{Do you want to overwrite out_file if it exists? Defaults to FALSE}
}
\value{
data frame or netcdf file.
}
\description{
Clip a netcdf file to a specific latlon box
}
\examples{

# Clipping data from the NOAA World Ocean Atlas sample file.
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")

# clip to a specific depth range using vert_range

nc_clip(ff, vert_range = c(1, 5))

# clip to a specific longitude and latitude box
# Clipping to the region around the UK

uk_sst <- nc_clip(ff, lon_range = c(-12, 10), lat_range = c(48, 62))
}
