% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{nc_dates}
\alias{nc_dates}
\title{Find dates in a netcdf file}
\usage{
nc_dates(ff)
}
\arguments{
\item{ff}{This is the file to read dates from.}
}
\description{
Find dates in a netcdf file
}
\examples{

# Get the dates in the sample file
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
nc_dates(ff)
}
