% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depths.R
\name{nc_depths}
\alias{nc_depths}
\title{get the depths available in a netcdf file}
\usage{
nc_depths(ff)
}
\arguments{
\item{ff}{This is the file to check.}
}
\description{
This function will tell you the depths available in a netcdf file
}
