% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_dir.R
\name{nc_dir}
\alias{nc_dir}
\title{search netcdf files in folder and list variables in each file}
\usage{
nc_dir(directory = getwd(), recurse = TRUE, print = FALSE)
}
\arguments{
\item{directory}{This is the directory to check.}

\item{recurse}{Do you want to recursive look through a directory?}

\item{print}{Do you want to print the name of files as you check them?}
}
\value{
data frame. First column is the file name. Remaining column are the variables, with TRUE signifying the variable is available in the file on that row.
}
\description{
This function will list the netcdf files in a directory that contain a specified variable
}
