% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{nc_merge}
\alias{nc_merge}
\title{merge netcdf files}
\usage{
nc_merge(ff_list, merge = "merge", expr = NULL, out_file = NULL,
  zip_file = FALSE, cdo_output = TRUE, overwrite = FALSE)
}
\arguments{
\item{ff_list}{This is a character vector containing the files to merge}

\item{merge}{This is the merge type. Use "merge" to combine files, "mergetime" to merge based on time.}

\item{expr}{This is a cdo expression to apply to the merged files.}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{zip_file}{Do you want any output file to be zipped to save space. Default is FALSE.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}

\item{overwrite}{Do you want to overwrite out_file if it exists? Defaults to FALSE}
}
\value{
data frame or netcdf file.
}
\description{
This function allows you to merge a list of netcdf files
}
\examples{
ff1 <- system.file("extdata", "icec.mon.ltm.1981-2010.nc", package = "rcdo")
ff2 <- system.file("extdata", "sst.mon.ltm.1981-2010.nc", package = "rcdo")
ff_list <- c(ff1, ff2)
coords <- expand.grid(Longitude = -39.5:10.5, Latitude = 40.5:69.5)
nc_merge(ff_list, cdo_output = TRUE)
}
