% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{nc_read}
\alias{nc_read}
\title{read ncdf file to a data frame}
\usage{
nc_read(ff, vars = NULL, date_range = NULL, cdo_output = FALSE)
}
\arguments{
\item{ff}{This is the file to move. This must be the full system path to the file.}

\item{vars}{A list of variables you want to read in. Character vector. Everything is read in if this is empty.}

\item{date_range}{This is the range of dates you want. c(date_min, date_max). "day/month/year" character string format.}

\item{cdo_output}{Do you want to show the cdo output? Set to TRUE in case you want to troubleshoot errors.}
}
\description{
This function allows you to clip a netcdf file to a specific latlon box
}
