% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{nc_read}
\alias{nc_read}
\title{Read a netcdf file to a data frame}
\usage{
nc_read(ff, vars = NULL, date_range = NULL, cdo_output = FALSE,
  dim_check = 1.5e+08)
}
\arguments{
\item{ff}{This is the file to read.}

\item{vars}{A list of variables you want to read in. Character vector. Everything is read in if this is empty.}

\item{date_range}{This is the range of dates you want. c(date_min, date_max). "day/month/year" character string format.}

\item{cdo_output}{Do you want to show the cdo output? Set to TRUE in case you want to troubleshoot errors.}

\item{dim_check}{The number of data points in the final data frame that will ask to continue. Set to NULL if you don't want to check.}
}
\description{
This is a quick and easy way to read a netcdf file to a data frame. It will read all or a specified list of variables into a data frame. You are able to specify a date range. This requires that grid details are clear in the netcdf file, so in some rare cases there will be an error message.
}
\examples{

# Reading in data from the NOAA World Ocean Atlas sample file.
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
# if we simply want to read the data into a tibble, we just need to use nc_read

nc_read(ff)

# By default nc_read reads in all data fields. But we probably just want to subset it
# If we only want to read in specific fields, we can use vars

nc_read(ff, vars = "t_an")
}
