% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{nc_remap}
\alias{nc_remap}
\title{remap a ncdf file}
\usage{
nc_remap(ff, vars = NULL, lon_range, lat_range, coord_res,
  date_range = NULL, months = NULL, years = NULL, out_file = NULL,
  vert_depths = NULL, remapping = "bil", cdo_output = FALSE)
}
\arguments{
\item{ff}{This is the file to move. This must be the full system path to the file.}

\item{vars}{Select the variables you want to regrid. If this is not given, all variables will be regridded.}

\item{lon_range}{longitude range. c(min_longituse, max_longitude).}

\item{lat_range}{latitude range. c(min_latitude, max_latitude).}

\item{date_range}{This is the range of dates you want. c(date_min, date_max). "day/month/year" character string format.}

\item{months}{Months you want. c(month_1, month_2,...)}

\item{years}{Months you want. c(year_1, year_2,...)}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{remapping}{The type of remapping. bil = bilinear. nn = nearest neighbour. dis = distance weighted.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}

\item{coord_rds}{longitudinal and latitudinal range for the regridding. c(lon_res, lat_res). If one number given the lon_res and lat_res will be the same.}
}
\value{
data frame or netcdf file.
}
\description{
This function allows you to remap a netcdf horizontally and vertically to a specific latlon box
}
\examples{
# Remapping NOAA world ocean atlas data to the region around the UK
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
# remapping to 1 degree resolution across all depth layers
nc_remap(ff, vars = "t_an", lon_range = c(-20, 10), lat_range = c(48,62), coord_res = 1)
# remapping to 1 degree resolution for 5, 50 and 100 metres
nc_remap(ff, vars = "t_an", lon_range = c(-20, 10), lat_range = c(48,62), vert_depths= c(5, 50, 100), coord_res = 1)

}
