% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{nc_variables}
\alias{nc_variables}
\title{List the variables in a netcdf file.}
\usage{
nc_variables(ff, detailed = FALSE)
}
\arguments{
\item{ff}{This is the file to read variables from.}

\item{details}{Do you want a data frame that provides the long name and units of the variable?}
}
\value{
list of variables
}
\description{
This function will tell you the variables in a netcdf file.
}
\examples{
# Get the variable names in the NOAA WOA sample file
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
nc_variables(ff)
# A data frame showing the variables, their long names and units
nc_variables(ff, details = TRUE)
}
