% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlimits.R, R/names.R
\name{nc_variables}
\alias{nc_variables}
\title{nc_variables}
\usage{
nc_variables(ff)

nc_variables(ff)
}
\arguments{
\item{ff}{This is the file to check. This must be the full system path to the file.}

\item{ff}{This is the file to read variables from.}
}
\description{
This function will tell you the variables in a netcdf file

This function will tell you the variables in a netcdf file
}
