% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertmean.R
\name{nc_vertmean}
\alias{nc_vertmean}
\title{Calculate vertical mean from a netcdf file.}
\usage{
nc_vertmean(ff, vars = NULL, vert_scale = NULL, coords = NULL,
  expr = NULL, na_value = NULL, out_file = NULL,
  cdo_output = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{ff}{This is the file to analyze.}

\item{vars}{Select the variables you want to regrid. If this is not given, all variables will be regridded.}

\item{vert_scale}{This is the vertical scale you want. c(min_depth, max_depth, vertical_resolution).}

\item{coords}{A 2 column matrix or data frame of the form (longitude, latitude) with coordinates for regridding. This can be regular or irregular. The function will calculate which it is.}

\item{expr}{This is a cdo expression to apply to the data before a vertical mean is calculated.}

\item{na_value}{This is a value in the raw netcdf file that needs to be treated as an na.}

\item{out_file}{The name of the file output. If this is not stated, a data frame will be the output.}

\item{cdo_output}{set to TRUE if you want to see the cdo output}

\item{overwrite}{Do you want to overwrite out_file if it exists? Defaults to FALSE}

\item{...}{Additional terms to be sent to nc_remap2}
}
\value{
data frame or netcdf file.
}
\description{
This function allows you to remap a netcdf horizontally and vertically to a specific latlon box. It outputs a data frame with the mean value and the maximum depth used for the calculation.
}
\examples{

# Calculating vertical mean NOAA world ocean atlas data to the region around the UK
ff <- system.file("extdata", "woa18_decav_t01_01.nc", package = "rcdo")
# calculate vertical mean of temperature between 5 and 30 m, using a resolution of 5 for vertical interpolation
nc_vertmean(ff, vars = "t_an", vert_scale = c(5, 30, 5))
}
