% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-ninfo.R
\name{ninfo}
\alias{ninfo}
\title{Print the number of parameters, levels or times}
\description{
This module prints the number of variables, levels or times of the input dataset.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{npar         Number of parameters
             Prints the number of parameters (variables).
nlevel       Number of levels
             Prints the number of levels for each variable.
nyear        Number of years
             Prints the number of different years.
nmon         Number of months
             Prints the number of different combinations of years and months.
ndate        Number of dates
             Prints the number of different dates.
ntime        Number of timesteps
             Prints the number of timesteps.
ngridpoints  Number of gridpoints
             Prints the number of gridpoints for each variable.
ngrids       Number of horizontal grids
             Prints the number of horizontal grids.
}\if{html}{\out{</div>}}
}
