% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-outputgmt.R, R/op-cdo-gmtcells.R,
%   R/op-cdo-gmtxyz.R
\name{outputgmt}
\alias{outputgmt}
\alias{cdo_gmtcells}
\alias{cdo_gmtxyz}
\title{GMT output}
\usage{
cdo_gmtcells(ifile)

cdo_gmtxyz(ifile)
}
\arguments{
\item{ifile}{String with the path to the input file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This module prints the first field of the input dataset to standard output. The output can be used to generate 2D Lon/Lat plots with GMT. The format of the output depends on the chosen operator.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{gmtxyz    GMT xyz format
          The operator exports the first field to the GMT xyz ASCII format.
          The output can be used to create contour plots with the GMT module pscontour.
gmtcells  GMT multiple segment format
          The operator exports the first field to the GMT multiple segment ASCII format.
          The output can be used to create shaded gridfill plots with the GMT module psxy.
}\if{html}{\out{</div>}}
}
