% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdo-package.R
\docType{package}
\name{rcdo-package}
\alias{rcdo}
\alias{rcdo-package}
\title{rcdo: Tools to manipulate netcdf data}
\description{
A package to manipulate netcdf files.
}
\author{
\strong{Maintainer}: Robert Wilson \email{robertwilson190@gmail.com}

}
\keyword{internal}
