% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-replace.R, R/op-cdo-replace.R
\name{replace}
\alias{replace}
\alias{cdo_replace}
\title{Replace variables}
\usage{
cdo_replace(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This operator replaces variables in infile1 by variables from infile2 and write the result to outfile. Both input datasets need to have the same number of timesteps. All variable names may only occur once!
}
