% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setrtoc.R, R/cdo-setrtoc2.R,
%   R/cdo-setvals.R, R/family-replacevalues.R
\name{cdo_setrtoc}
\alias{cdo_setrtoc}
\alias{cdo_setrtoc2}
\alias{cdo_setvals}
\alias{replacevalues}
\title{Replace variable values}
\usage{
cdo_setrtoc(
  ifile,
  oldval,
  newval,
  ... = NULL,
  rmin = NULL,
  rmax = NULL,
  c = NULL,
  c2 = NULL,
  ofile = NULL
)

cdo_setrtoc2(
  ifile,
  oldval,
  newval,
  ... = NULL,
  rmin = NULL,
  rmax = NULL,
  c = NULL,
  c2 = NULL,
  ofile = NULL
)

cdo_setvals(
  ifile,
  oldval,
  newval,
  ... = NULL,
  rmin = NULL,
  rmax = NULL,
  c = NULL,
  c2 = NULL,
  ofile = NULL
)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{oldval, newval, ..., }{FLOAT - Pairs of old and new values}

\item{rmin, }{FLOAT - Lower bound}

\item{rmax, }{FLOAT - Upper bound}

\item{c, }{FLOAT - New value - inside range}

\item{c2, }{FLOAT - New value - outside range}

\item{ofile}{String with the path to the output file.}
}
\description{
This module replaces old variable values with new values, depending on the operator.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{setvals   Set list of old values to new values
          Supply a list of n pairs of old and new values.
setrtoc   Set range to constant
                   / c      if i(t,x) GE rmin AND i(t,x) LE rmax
          o(t,x) = 
                   \\\\ i(t,x) if i(t,x) LT rmin AND i(t,x) GT rmax
setrtoc2  Set range to constant others to constant2
                   / c      if i(t,x) GE rmin AND i(t,x) LE rmax
          o(t,x) = 
                   \\\\ c2     if i(t,x) LT rmin AND i(t,x) GT rmax
}\if{html}{\out{</div>}}
}
