% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-setgridcell.R, R/family-setgridcell.R
\name{cdo_setgridcell}
\alias{cdo_setgridcell}
\alias{setgridcell}
\title{Set the value of a grid cell}
\usage{
cdo_setgridcell(ifile, value = NULL, cell = NULL, mask = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{value}{FLOAT - Value of the grid cell}

\item{cell}{INTEGER - Comma-separated list of grid cell indices}

\item{mask}{STRING - Name of the data file which contains the mask}

\item{ofile}{String with the path to the output file.}
}
\description{
This operator sets the value of the selected grid cells. The grid cells can be selected by a comma-separated list of grid cell indices or a mask. The mask is read from a data file, which may contain only one field. If no grid cells are selected, all values are set.
}
