% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-genlevelbounds.R, R/cdo-setzaxis.R,
%   R/family-setzaxis.R
\name{cdo_genlevelbounds}
\alias{cdo_genlevelbounds}
\alias{cdo_setzaxis}
\alias{setzaxis}
\title{Set zaxis information}
\usage{
cdo_genlevelbounds(ifile, zaxis = NULL, zbot = NULL, ztop = NULL, ofile = NULL)

cdo_setzaxis(ifile, zaxis = NULL, zbot = NULL, ztop = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{zaxis}{STRING - Z-axis description file or name of the target z-axis}

\item{zbot}{FLOAT - Specifying the bottom of the vertical column. Must have the same units as z-axis.}

\item{ztop}{FLOAT - Specifying the top of the vertical column. Must have the same units as z-axis.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
This module modifies the metadata of the vertical grid.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{setzaxis        Set z-axis
                This operator sets the z-axis description of all variables with the same number of level as the new z-axis.
genlevelbounds  Generate level bounds
                Generates the layer bounds of the z-axis.
}\if{html}{\out{</div>}}
}
