% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-sp2sp.R, R/family-specconv.R
\name{cdo_sp2sp}
\alias{cdo_sp2sp}
\alias{specconv}
\title{Spectral conversion}
\usage{
cdo_sp2sp(ifile, trunc = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{trunc, }{INTEGER - New spectral resolution}

\item{ofile}{String with the path to the output file.}
}
\description{
Changed the triangular truncation of all spectral fields. This operator performs downward conversion by cutting the resolution. Upward conversions are achieved by filling in zeros.
}
