% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-splitdate.R
\name{splitdate}
\alias{splitdate}
\title{Splits a file into dates}
\description{
This operator splits infile into pieces, one for each different date. The output files will be named <obase><YYYY-MM-DD><suffix> where YYYY-MM-DD is the date and suffix is the filename extension derived from the file format.
}
