% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family-timcor.R, R/op-cdo-timcor.R
\name{timcor}
\alias{timcor}
\alias{cdo_timcor}
\title{Correlation over time}
\usage{
cdo_timcor(ifile1, ifile2, ofile = NULL)
}
\arguments{
\item{ifile1, ifile2}{Strings with the path to the input files.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operators that don't return filenames return a character vector with the
string output.
}
\description{
The correlation coefficient is a quantity that gives the quality of a least squares fitting to the original data. This operator correlates each gridpoint of two fields over all timesteps. If there is only one input field, the p-value (probability value) is also written out. With  S(x) = \{t, i_1(t,x) != missval and i_2(t,x) != missval\}  it is  o(1,x) = Cor\{(i_1(t,x), i_2(t,x)), t_1 < t <= t_n\}  For every gridpoint x only those timesteps t belong to the sample, which have i_1(t,x) != missval and i_2(t,x) != missval.
}
