% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-timcumsum.R, R/family-timcumsum.R
\name{cdo_timcumsum}
\alias{cdo_timcumsum}
\alias{timcumsum}
\title{Cumulative sum over all timesteps}
\usage{
cdo_timcumsum(ifile, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{ofile}{String with the path to the output file.}
}
\value{
Operators that output one or more files return a character vector to the
output files.

Operators that output an indefinite number of files return a string with the
basename of the files.

Operatos that don't return filenames return a character vector with the
string output.
}
\description{
The timcumsum operator calculates the cumulative sum over all timesteps. Missing values are treated as numeric zero when summing.  o(t,x) = sum\{i(t',x), 0<t'<=t\}
}
