% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo-gh2hl.R, R/family-vertintgh.R
\name{cdo_gh2hl}
\alias{cdo_gh2hl}
\alias{vertintgh}
\title{Vertical height interpolation}
\usage{
cdo_gh2hl(ifile, hlevels = NULL, ofile = NULL)
}
\arguments{
\item{ifile}{String with the path to the input file.}

\item{hlevels, }{FLOAT - Comma-separated list of height levels in meter}

\item{ofile}{String with the path to the output file.}
}
\description{
Interpolate 3D variables on hybrid sigma height coordinates to height levels. The input file must contain the 3D geometric height in meter. The geometric height is identified by the NetCDF CF standard name geometric_height_at_full_level_center. Use the alias  gh2hlx or the environment variable EXTRAPOLATE to extrapolate missing values. This operator requires all variables on the same horizontal grid.
}
\section{Note}{

This is a specific implementation for NetCDF files from the ICON model, it may not work with data from other sources.
}

